refactor(testpage): move MockTestPage.cs's 16 non-LiveNavTestPage types into six siblings
Pure move at top-level type boundaries: no declaration edits, no logic. 41 build
warnings before and after.
