+++ AlRunner/Patches/MockTestPage.cs
@@ -18,173 +18,6 @@ using Microsoft.Dynamics.Nav.Types.Exceptions;
 
 namespace AlRunner;
 
-/// <summary>
-/// The field numbers of a record's current key, for <c>ITestFilter.GetCurrentKeyFields</c>
-/// (#3316). Shared by the two implementations that back a filter with a real NavRecord —
-/// <see cref="LiveNavTestPage"/> and RequestPageTestPage's data-item filter — so the two
-/// cannot answer the same question differently.
-/// </summary>
-internal static class TestFilterKeyFields
-{
-    /// <summary>
-    /// NCLMetaKey's field list is internal and NavRecord exposes only ALCurrentKey (the key
-    /// rendered by name) and ALCurrentKeyIndex, so the numbers are recovered by resolving each
-    /// rendered field name against the table's own metadata. A name the table does not know is
-    /// skipped rather than guessed at: an invented field number would be a wrong answer, where
-    /// a short list is a visibly incomplete one.
-    /// </summary>
-    internal static int[] Of(NavRecord record)
-    {
-        var names = SplitRenderedKey(record.ALCurrentKey);
-        if (names.Length == 0) return Array.Empty<int>();
-
-        var nos = new List<int>();
-        foreach (var name in names)
-            foreach (var field in record.MetaTable.Fields)
-            {
-                if (!string.Equals(field.FieldName, name, StringComparison.OrdinalIgnoreCase)) continue;
-                nos.Add(field.FieldNo);
-                break;
-            }
-        return nos.ToArray();
-    }
-
-    /// <summary>
-    /// The field names out of a key as NavRecord.ALCurrentKey renders it — comma-separated,
-    /// each name optionally quoted because BC quotes any name that is not a bare identifier
-    /// ("Entry No." carries a space and a period). Split out from <see cref="Of"/> so the
-    /// parsing can be asserted without a live NavRecord.
-    /// </summary>
-    internal static string[] SplitRenderedKey(string? rendered)
-    {
-        if (string.IsNullOrEmpty(rendered)) return Array.Empty<string>();
-
-        var names = new List<string>();
-        foreach (var part in rendered!.Split(','))
-        {
-            var name = part.Trim().Trim('"');
-            if (name.Length != 0) names.Add(name);
-        }
-        return names.ToArray();
-    }
-}
-
-/// <summary>
-/// Minimal ITestPage + ITestFilter + IDisposable implementation.
-/// All field/action/filter state is held in plain dictionaries; navigation
-/// always reports "no more rows" (returns false / empty).
-/// </summary>
-internal class MockITestPage : ITestPage
-{
-    private readonly Dictionary<int, string>      _filters     = new();
-    private readonly Dictionary<int, MockITestField>  _fields  = new();
-    private readonly Dictionary<int, MockITestAction> _actions = new();
-    private bool   _ascending        = true;
-    private int[]? _currentKeyFields;
-
-    /// <summary>
-    /// The refusals this PAGE's controls have recorded, read back by ValidationErrorCount /
-    /// GetValidationError below (#3009). Protected so LiveNavTestPage can hand it to every
-    /// control it builds — one ledger per page, shared by all of that page's controls, which
-    /// is the shape real BC has.
-    /// </summary>
-    private protected readonly TestPageValidationErrors _validationErrors = new();
-
-    // ── ITestPage ──────────────────────────────────────────────────────────
-
-    // IsOpened() = false so NavTestPageBase.Open() "already open" guard passes.
-    public virtual bool IsOpened()  => false;
-    public virtual void Close()     { }
-    public virtual void Dispose()   { }
-
-    public virtual ITestField GetField(int id)
-    {
-        if (!_fields.TryGetValue(id, out var f))
-            _fields[id] = f = new MockITestField();
-        return f;
-    }
-
-    public virtual ITestAction GetAction(int id)
-    {
-        if (!_actions.TryGetValue(id, out var a))
-            _actions[id] = a = new MockITestAction();
-        return a;
-    }
-
-    public virtual ITestPart  GetPart(int id)                                           => new MockITestPart();
-    public virtual ITestAction GetBuiltInAction(FormResult formResult)                  => new MockITestAction();
-    public virtual ITestFilter GetDataItemFilter(string id)                              => this;
-    public void               SetSelection(bool value)                                  { }
-    public virtual void       InsertEmptyRow(bool beforeCurrent)                        { }
-    public virtual bool       MoveNext()                                                => false;
-    public virtual bool       MovePrevious()                                            => false;
-    public virtual bool       MoveFirst()                                               => false;
-    public virtual bool       MoveLast()                                                => false;
-    // #3009: was hardcoded `string.Empty`, so a validation error real BC reports on the PAGE
-    // came back blank. Backed by a real ledger now; the base mock simply never has anything in
-    // it, because a page with no live instance has no controls that can record a refusal. Not
-    // virtual: LiveNavTestPage feeds the SAME ledger rather than overriding, which is what
-    // stops the two views from drifting. See TestPageValidationErrors.
-    public string             GetValidationError(int index)                             => _validationErrors.Get(index);
-    public virtual bool       FindRowFromTableFieldValues(int[] f, object[] v, bool fw) => false;
-    public virtual bool       FindRowFromControlFieldValue(int fId, object v, bool fw)  => false;
-    public virtual object?    GetBookmark()                                             => null;
-    public virtual bool       GoToBookmark(object bookmark)                             => false;
-    public virtual object[]   GetTableFieldValues(int[] fieldIds)                       => Array.Empty<object>();
-    // Virtual since #3185: LiveNavTestPage answers these with the page's real built-in
-    // page-mode actions. The base mock keeps the no-op action a page with no live instance
-    // has always had.
-    public virtual ITestAction Edit()                                                   => new MockITestAction();
-    public virtual ITestAction View()                                                   => new MockITestAction();
-    public bool               Expand(bool doExpand)                                     => false;
-
-    // #3009: was hardcoded 0. BC's NavTestPageBase.ALValidationErrorCount() reads this member
-    // straight through to AL's TestPage.ValidationErrorCount(), and NavTestField.CheckError
-    // reads it before and after every control write, so a constant 0 made the page branch of
-    // CheckError unreachable and made AL's own read answer 0 after a refusal BC counts.
-    public int        ValidationErrorCount => _validationErrors.Count;
-    public virtual FormResult FormResult   => FormResult.OK;
-    public string     Name                 => string.Empty;
-    public virtual string Caption          => string.Empty;
-    public virtual int PageId            => 0;
-    public virtual Guid FormHandle         => Guid.Empty;
-    public virtual bool Creatable          => false;
-    public bool       IsExpanded           => false;
-    public virtual bool RuntimeEditable    => true;
-
-    // ── ITestFilter (inherited via ITestPage) ─────────────────────────────
-
-    public virtual void SetFilter(int fieldId, string filterValue) => _filters[fieldId] = filterValue;
-    public IEnumerable<NavFilter> GetFilter() => Array.Empty<NavFilter>();
-    public virtual string GetFilter(int fieldId) => _filters.TryGetValue(fieldId, out var v) ? v : string.Empty;
-    // Virtual since #3316: a page with a live rowset holds its key and direction on the
-    // NavRecord it walks, not here — see LiveNavTestPage's overrides. This base implementation
-    // is what a page with no record has: it remembers what was set, because nothing else can.
-    public virtual void SetCurrentKeyFields(int[] fields) { _currentKeyFields = fields; }
-    public virtual int[]  GetCurrentKeyFields() => _currentKeyFields ?? Array.Empty<int>();
-
-    public virtual bool   Ascending
-    {
-        get => _ascending;
-        set => _ascending = value;
-    }
-
-    // #3316: this used to join the raw field NUMBERS, so AL's TestPage.Filter.CurrentKey()
-    // answered "2, 3" where BC answers a key naming its fields. NavTestFilter.ALCurrentKey
-    // reads this member straight through, so the rendering is wholly ours. Without a record
-    // there is no metadata to resolve a number against, so a recordless page renders the
-    // numbers as before rather than inventing names; LiveNavTestPage overrides with the real
-    // one. See docs/limitations.md.
-    public virtual string CurrentKey
-    {
-        get
-        {
-            if (_currentKeyFields == null || _currentKeyFields.Length == 0) return string.Empty;
-            return string.Join(", ", _currentKeyFields);
-        }
-    }
-}
-
 internal class LiveNavTestPage : MockITestPage
 {
     // Null for a page with no SourceTable (issue #2007) — a legal AL shape (StandardDialog
@@ -1016,8 +849,6 @@ internal class LiveNavTestPage : MockITestPage
     // Loaded() (the record-positioning trigger) sets this flag.
     private bool _tornDown;
 
-
-
     // Set only around the page-construction-time initial positioning call (MarkOpened /
     // RunnerTestClientSession.GetPage's own MoveFirst()). MarkOpened's caller wraps it in a
     // blanket `catch { }` that would swallow whatever Loaded() throws there; GetPage's is not
@@ -1212,7 +1043,6 @@ internal class LiveNavTestPage : MockITestPage
     // dropped every insert made through a TestPage; a LIVE page has a real record, so it
     // must initialise the buffer and remember to flush it.
     private bool _pendingNewRow;
-
 
     /// <summary>
     /// Turn the current position into a pending insert.
@@ -2893,7 +2723,6 @@ internal class LiveNavTestPage : MockITestPage
     /// </summary>
     public override string CurrentKey => _record == null ? string.Empty : _record.ALCurrentKey;
 
-
     /// <summary>
     /// Resolve a CONTROL id to the source-table field it is bound to. A control bound to a
     /// page variable is not in the rowset and cannot be used to locate a row, so this
@@ -2930,2069 +2759,4 @@ internal class LiveNavTestPage : MockITestPage
         right = Unwrap(right);
         return Equals(left, right);
     }
-}
-
-/// <summary>
-/// Option values as a TestPage sees them: member NAMES going in, a member name coming back out.
-///
-/// AL's TestPage API is string-typed for every control — <c>Field.SetValue('Sum')</c>,
-/// <c>Field.Value()</c> — so the option's member table is the only thing that can turn that
-/// string into the ordinal the record stores, and back. Without it a write puts a NavText into
-/// an Option and dies inside BC's own setter ("The value \"Sum\" can't be evaluated into type
-/// Option"), and a read answers with the bare ordinal, which no AL test is written against.
-///
-/// Shared by the Rec-bound field and the page-variable-bound field. It was originally written
-/// for the latter only, which is exactly the shape of bug worth avoiding here: the two kinds of
-/// control look identical in AL, so a test author has no way to know that one of them resolves
-/// option names and the other does not.
-/// </summary>
-internal static class TestPageOptionValue
-{
-    /// <summary>Turn the string a test wrote into the NavOption the binding holds.</summary>
-    internal static NavValue Resolve(NavOption current, string value, string[]? captions, string context)
-    {
-        var metadata = current.NavOptionMetadata
-            ?? throw TestPageShapeGap.OptionValue(
-                context,
-                "the control is bound to an Option with no option metadata, so a value cannot "
-                + "be resolved by name");
-
-        var options = Members(metadata);
-        var ordinals = Ordinals(metadata);
-
-        // A TestPage sets an option by what the user sees, i.e. the control's OptionCaption,
-        // which is NOT the option's member names (Pageworks: captions
-        // "Fields,Blocks,Images,…" over members [Field, Block, Image, …]). Captions first,
-        // then members — the caption is what AL test code is written against.
-        if (captions != null)
-            for (var i = 0; i < captions.Length; i++)
-                if (OptionNamesEqual(captions[i], value))
-                    return NavOption.Create(metadata, OrdinalAt(ordinals, i));
-
-        // Issue #1928, decided against real-BC evidence (StefanMaron/BusinessCentral.AL.
-        // Language.Tests#50, run against a real BC service tier on two BC versions): an
-        // Enum-typed control's TestPage.SetValue resolves ONLY by the declared Caption and
-        // REFUSES the member name — SetValue('Block') against `value(1; Block) { Caption =
-        // 'Blocks'; }` throws "Your entry of 'Block' is not an acceptable value for
-        // 'Kind'.", not a successful set. So for an Enum-backed metadata (IsEnum), the
-        // member-name fallback below must NOT run — accepting a spelling real BC rejects is
-        // exactly the silent divergence loud-failures.md forbids, and it is what shipped as
-        // a ghost test in tests/runner-extras/page-enum-control-modal before this fix.
-        //
-        // The plain `Option` primitive is a SEPARATE, unverified question — no real-BC
-        // evidence either way distinguishes caption-vs-member resolution for it, so its
-        // historical member-name fallback stays as-is; only Enum's is removed here.
-        var isEnumBacked = metadata.IsEnum;
-        if (!isEnumBacked)
-            for (var i = 0; i < options.Length; i++)
-                if (OptionNamesEqual(options[i], value))
-                    return NavOption.Create(metadata, OrdinalAt(ordinals, i));
-
-        // A bare number is a legal way to set an option, and unambiguous.
-        if (int.TryParse(value, System.Globalization.NumberStyles.Integer,
-                CultureInfo.InvariantCulture, out var literal)
-            && (ordinals == null || ordinals.Contains(literal)))
-            return NavOption.Create(metadata, literal);
-
-        throw new AlRunner.Infrastructure.RunnerOutOfScopeException(
-            context,
-            isEnumBacked
-                ? $"testpage-option-value — '{value}' is not an acceptable value. An "
-                  + "Enum-typed control resolves TestPage.SetValue by its declared Caption "
-                  + "only, never by the member name (real BC's own behavior — see issue "
-                  + "#1928) — "
-                  + (captions != null
-                      ? $"acceptable captions are [{string.Join(", ", captions)}]"
-                      : "the enum declares no captions")
-                  + $". Member names ([{string.Join(", ", options)}]) are NOT accepted. "
-                  + "See docs/scope.md"
-                : $"testpage-option-value — '{value}' is not one of the option's values "
-                  + $"[{string.Join(", ", options)}]"
-                  + (captions != null
-                      ? $" nor one of its captions [{string.Join(", ", captions)}]"
-                      : " (the control declares no OptionCaption)")
-                  + ". See docs/scope.md");
-    }
-
-    /// <summary>
-    /// The text a test reads back. Deliberately the same spelling <see cref="Resolve"/> accepts
-    /// first, so <c>SetValue(Value())</c> is a no-op — a page whose read and write disagreed
-    /// about captions-vs-members would let a test copy a value from one field to another and
-    /// silently write a different member.
-    /// </summary>
-    internal static string? Display(NavOption option, string[]? captions)
-        => option.NavOptionMetadata is { } metadata
-            ? DisplayOrdinal(metadata, option.Value, captions)
-            : null;
-
-    /// <summary>
-    /// The same text <see cref="Display"/> produces, for a BARE ORDINAL rather than for a
-    /// NavOption that already carries its own metadata (issue #2367).
-    ///
-    /// <c>NavTestField.ALAssertEquals</c> and <c>ALSetValue</c> — the real, precompiled BC
-    /// methods the AL compiler emits for <c>TestPage.&lt;field&gt;.AssertEquals(&lt;option&gt;)</c>
-    /// and <c>SetValue(&lt;option&gt;)</c> — never hand an AL Option/Enum value to
-    /// <see cref="ITestField"/> as-is. They round-trip it through
-    /// <c>NavValue.CreateNavValueFromObject(NavValueMetadata.DefaultMetadata(FieldType), value)</c>,
-    /// whose <c>NavNclType.NavOption</c> arm rebuilds the value against the DEFAULT option
-    /// metadata, and then hand the resulting <c>ClientObject</c> — a bare ordinal, with the
-    /// field's own member/caption table gone — to <see cref="ITestField.ValueToString"/>.
-    ///
-    /// So <c>ValueToString</c> is where the control's own option table has to be put back.
-    /// The metadata comes from the value the control currently holds, which is the control's
-    /// option set; only the ordinal comes from the caller.
-    /// </summary>
-    internal static string? DisplayOrdinal(NavOption? current, object? value, string[]? captions)
-        => current?.NavOptionMetadata is { } metadata && TryAsOrdinal(value, out var ordinal)
-            ? DisplayOrdinal(metadata, ordinal, captions)
-            : null;
-
-    private static string? DisplayOrdinal(object metadata, int ordinal, string[]? captions)
-    {
-        var index = IndexOfOrdinal(metadata, ordinal);
-        if (index < 0) return null;
-
-        if (captions != null && index < captions.Length) return captions[index];
-        var options = Members(metadata);
-        return index < options.Length ? options[index] : null;
-    }
-
-    // Deliberately narrow. An ordinal is what BC's own round trip leaves behind for an
-    // Option/Enum, and nothing else on this path is one: a string is not accepted here
-    // because ALSetValue's `value is NavStringValue` fast path never reaches ValueToString,
-    // so a string arriving would mean some OTHER caller with an unexamined contract, and
-    // silently reinterpreting its text as an option would be exactly the kind of guess
-    // loud-failures.md exists to prevent. Anything unrecognised falls back to the caller's
-    // own Convert.ToString, i.e. to the behaviour before #2367.
-    private static bool TryAsOrdinal(object? value, out int ordinal)
-    {
-        switch (value)
-        {
-            case NavOption option: ordinal = option.Value; return true;
-            case int i:            ordinal = i;            return true;
-            case short s:          ordinal = s;            return true;
-            case byte b:           ordinal = b;            return true;
-            case sbyte sb:         ordinal = sb;           return true;
-            case long l when l >= int.MinValue && l <= int.MaxValue:
-                                   ordinal = (int)l;       return true;
-            default:               ordinal = 0;            return false;
-        }
-    }
-
-    /// <summary>
-    /// An Enum-typed control's per-value captions, sourced from the enum's OWN metadata.
-    ///
-    /// Unlike the <c>Option</c> primitive, an AL <c>Enum</c> has no page-level
-    /// <c>OptionCaption</c> property to declare, so
-    /// <see cref="AlRunner.Patches.RunnerPageInstance.TryGetOptionCaptions"/>'s
-    /// <c>ControlDefinition.OptionCaptionML</c> lookup is always empty for it (verified via
-    /// <c>AL_RUNNER_TRACE_PAGE_METADATA=2</c> against an Enum-bound page-variable control:
-    /// <c>OptionCaption='' OptionCaptionML=''</c>). Real BC computes an Enum's captions
-    /// from its own metadata instead — see issue #1928's real-BC evidence: a real service
-    /// tier's <c>TestPage.SetValue</c> on an Enum control resolves by the declared
-    /// <c>Caption</c> and REFUSES the member name (the exact opposite of what this runner
-    /// did before this fix).
-    ///
-    /// <c>IsEnum</c>/<c>GetOrdinals()</c>/<c>GetCaptionFromIndex(int)</c> are public virtuals
-    /// on <c>NCLOptionMetadata</c> (decompiled: <c>Microsoft.Dynamics.Nav.Ncl.dll</c>), which
-    /// <c>AlEnumOptionMetadata</c> (EnumMetadataPatches.cs) overrides from the SAME
-    /// emit-captured <c>(name, options[], indexes[], captions[])</c> tuple already used, and
-    /// already accepted as faithful, for <c>Enum::"X".Ordinals()/.Names()</c> via
-    /// <c>NCLEnumMetadata_CreateByIdAlAware</c>. The result is built in
-    /// <c>GetOrdinals()</c> order, which is the SAME order <see cref="Ordinals"/>'s reflection
-    /// (over a different, private accessor) already returns for the same metadata instance —
-    /// both walk the one <c>(options[], indexes[])</c> pair the AL emit captured — so a
-    /// caption at index i here lines up with the member at index i in <see cref="Members"/>,
-    /// which is what <see cref="Resolve"/> and <see cref="Display"/> index into.
-    ///
-    /// Returns null for a plain <c>Option</c> value (<c>IsEnum</c> is false there) or when
-    /// no bound value is available — the caller falls back to member-name display/resolution,
-    /// same as when a control declares no <c>OptionCaption</c> at all.
-    /// </summary>
-    internal static string[]? EnumCaptions(NavOption? option)
-    {
-        if (option?.NavOptionMetadata is not { IsEnum: true } metadata) return null;
-
-        var ordinals = new System.Collections.Generic.List<int>();
-        foreach (var ordinal in metadata.GetOrdinals()) ordinals.Add(ordinal);
-
-        var captions = new string[ordinals.Count];
-        for (var i = 0; i < ordinals.Count; i++)
-            captions[i] = metadata.GetCaptionFromIndex(ordinals[i]);
-        return captions;
-    }
-
-    /// <summary>The number of members, for AL that walks an option set rather than naming one.</summary>
-    internal static int Count(NavOption option)
-        => option.NavOptionMetadata is { } metadata ? Members(metadata).Length : 0;
-
-    /// <summary>The member at a position, in the same spelling <see cref="Display"/> uses.</summary>
-    internal static string MemberAt(NavOption option, int index, string[]? captions)
-    {
-        if (captions != null && index >= 0 && index < captions.Length) return captions[index];
-        if (option.NavOptionMetadata is not { } metadata) return string.Empty;
-        var options = Members(metadata);
-        return index >= 0 && index < options.Length ? options[index] : string.Empty;
-    }
-
-    // Options / OrdinalValues are internal to Ncl — read them reflectively rather than
-    // re-deriving the option set from OptionString, which would lose the ordinal gaps a
-    // declared option set is allowed to have.
-    private static string[] Members(object metadata)
-        => ReadNonPublic<string[]>(metadata, "Options") ?? Array.Empty<string>();
-
-    private static int[]? Ordinals(object metadata) => ReadNonPublic<int[]>(metadata, "OrdinalValues");
-
-    private static int OrdinalAt(int[]? ordinals, int index)
-        => ordinals != null && index < ordinals.Length ? ordinals[index] : index;
-
-    private static int IndexOfOrdinal(object metadata, int ordinal)
-    {
-        var ordinals = Ordinals(metadata);
-        if (ordinals == null)
-            return ordinal >= 0 && ordinal < Members(metadata).Length ? ordinal : -1;
-        return Array.IndexOf(ordinals, ordinal);
-    }
-
-    private static T? ReadNonPublic<T>(object target, string name) where T : class
-    {
-        for (var t = target.GetType(); t != null; t = t.BaseType)
-        {
-            var pi = t.GetProperty(name, System.Reflection.BindingFlags.Public
-                | System.Reflection.BindingFlags.NonPublic
-                | System.Reflection.BindingFlags.Instance
-                | System.Reflection.BindingFlags.DeclaredOnly);
-            if (pi != null) return pi.GetValue(target) as T;
-        }
-        return null;
-    }
-
-    // AL option names are compared ignoring case and spacing, the same way the runner
-    // compares object and field names elsewhere ("Custom Fields" vs "CustomFields").
-    private static bool OptionNamesEqual(string left, string right)
-        => string.Equals(left.Replace(" ", string.Empty), right.Replace(" ", string.Empty),
-            StringComparison.OrdinalIgnoreCase);
-}
-
-/// <summary>
-/// Boolean values as a TestPage sees them, on either shape of control: a page-variable-bound
-/// one (<c>field(Flag; ShowFlag)</c> where <c>ShowFlag: Boolean</c>) or a Rec-bound one
-/// (<c>field(Flag; Rec.Flag)</c> where the source table field is <c>Boolean</c>) — see issue
-/// #1870, the Rec-bound half of #1837 that #1869 (the page-variable half) left open.
-///
-/// <c>NavTestField.ALSetValue</c> — the real, precompiled BC method the AL compiler emits for
-/// every <c>TestPage.&lt;field&gt;.SetValue(&lt;Boolean&gt;)</c> call — never hands a NavValue
-/// straight to <see cref="ITestField"/>. For anything that is not itself already a
-/// <c>NavStringValue</c> it round-trips through <see cref="ITestField.FieldType"/> (to pick a
-/// <c>NavValueMetadata</c>) and then <see cref="ITestField.ValueToString"/> (both OUR OWN mock
-/// methods) to turn the boolean back into a string before ever reaching <see cref="ITestField.Value"/>'s
-/// setter — see the doc comment on <see cref="PageVariableTestField.FieldType"/> for why that
-/// matters here. <see cref="LiveNavTestField.FieldType"/> is sourced from the source table
-/// field's own declared type instead, but reaches the same <c>NavType.Boolean</c> answer for a
-/// <c>Boolean</c> field, so the round trip is identical on both sides.
-///
-/// Because both ends of that round trip are code THIS runner owns (<see cref="ITestField.ValueToString"/>
-/// always answers with <c>Convert.ToString(boolValue)</c>, i.e. exactly "True" or "False"), accepting
-/// only that spelling here is not a narrowing of what <c>SetValue(&lt;Boolean&gt;)</c> can express —
-/// it is the ONLY spelling that overload ever produces. Anything else (a literal
-/// <c>SetValue('Yes')</c>, locale spellings, ...) is a genuinely separate, upstream-unvalidated
-/// question about what real BC's own text-to-Boolean evaluate accepts on this surface, so it stays
-/// out of scope here and throws loudly rather than guessing.
-/// </summary>
-/// <summary>
-/// Enforces a field's declared <c>MinValue</c>/<c>MaxValue</c> AL properties on a TestPage
-/// control write (issue #2495). Measured against real BC (28.1 / 28.4, see #2490's arm A2):
-/// a Decimal field with <c>MinValue = 0;</c> raises
-/// <c>Validation error for Field: &lt;caption&gt;,  Message = 'The value must be greater than
-/// or equal to 0. Value: -1.00. (Select Refresh to discard errors)'</c> from a TestPage
-/// SetValue, while the SAME write via <c>Rec.Validate</c> or a plain field assignment raises
-/// nothing at all — this is a client/page-layer check, not a table-trigger one, so it must
-/// stay out of NavRecord.ALValidateAsync (which Rec.Validate also calls).
-///
-/// <para>Since #2900 this raises only the CORE of that message. Two layers are added around it
-/// by code that is not this helper's: <see cref="TestFieldValidationErrors"/> appends
-/// <c>" (Select Refresh to discard errors)"</c> when it records the refusal (BC's client does
-/// that, measured on corpus run 34002487601), and BC's own <c>NavTestField.CheckError</c> then
-/// wraps the result in <c>Validation error for Field: &lt;name&gt;,  Message = '…'</c>. The
-/// AL-visible string is the same one #2490 measured; it is now composed rather than
-/// assembled here, which is what stops the suffix appearing twice.</para>
-///
-/// <para>Only numeric field types are checked — MinValue/MaxValue is meaningless on Text/Code/
-/// Boolean/etc., and AL does not let those types declare it.</para>
-///
-/// <para>The bound text is read via <see cref="RecordPatches.TryGetParsedFieldMinMax"/> — the
-/// parse-time source, not the constructed <c>NCLMetaField</c> — because NCLMetaField
-/// (Microsoft.Dynamics.Nav.Runtime) does not expose MinValue/MaxValue on the built runtime
-/// object at all (confirmed empirically: no Min/MaxValue-named member of any accessibility),
-/// even though <c>MetaField</c> (Microsoft.Dynamics.Nav.Types.Metadata, what the runner's
-/// NclMetaTableBuilder constructs FROM) does carry them. Same rationale as
-/// <see cref="RecordPatches.TryGetParsedFieldCaption"/> reading Caption straight from the parsed
-/// table rather than through NCLMetaField's own getter.</para>
-/// </summary>
-internal static class TestPageMinMaxValue
-{
-    internal static void Check(NCLMetaTable table, int fieldNo, NavType fieldType, string rawValue, string caption)
-    {
-        if (fieldType is not (NavType.Decimal or NavType.Integer or NavType.BigInteger)) return;
-        if (!table.TryGetFieldByNo(fieldNo, out var field) || field == null) return;
-
-        var (minText, maxText) = RecordPatches.TryGetParsedFieldMinMax(table.TableId, fieldNo);
-        if (string.IsNullOrEmpty(minText) && string.IsNullOrEmpty(maxText)) return;
-
-        // AL's TestPage SetValue is string-typed for every control, and the generic
-        // ALCompiler.ToNavValue(value) path taken above (there is no per-type conversion for a
-        // Decimal/Integer control here, unlike the Boolean special-case) always produces a
-        // NavText — so the bound check parses the RAW string the test wrote, exactly as BC's own
-        // client-side field validation would before ever constructing a typed NavValue.
-        if (!decimal.TryParse(rawValue, NumberStyles.Any, CultureInfo.InvariantCulture, out var value)) return;
-
-        var isInteger = fieldType is NavType.Integer or NavType.BigInteger;
-
-        if (!string.IsNullOrEmpty(minText) && decimal.TryParse(minText, NumberStyles.Any, CultureInfo.InvariantCulture, out var min)
-            && value < min)
-            throw MakeError(isInteger, "greater than or equal to", minText!, value);
-
-        if (!string.IsNullOrEmpty(maxText) && decimal.TryParse(maxText, NumberStyles.Any, CultureInfo.InvariantCulture, out var max)
-            && value > max)
-            throw MakeError(isInteger, "less than or equal to", maxText!, value);
-    }
-
-    // The offending VALUE renders with the field's decimal places (2 by default for a Decimal,
-    // none for an Integer/BigInteger — measured against real BC, #2490's arm A2: "-1.00"). The
-    // BOUND is echoed as BC declared it (the raw MinValue/MaxValue AL text, e.g. "0"), NOT
-    // reformatted to the field's decimal places — also measured in #2490's arm A2, where the
-    // bound reads "0" while the value reads "-1.00" for the identical Decimal field.
-    private static string FormatValue(decimal d, bool isInteger)
-        => isInteger ? d.ToString("0", CultureInfo.InvariantCulture)
-                     : d.ToString("0.00", CultureInfo.InvariantCulture);
-
-    private static System.Exception MakeError(bool isInteger, string comparison, string boundText, decimal value)
-    {
-        // The BARE message only. BC's own NavTestField.CheckError wraps whatever an ITestField
-        // records into "Validation error for Field: {Name},  Message = '{recorded}'" using
-        // Lang.TestValidationException, so building that wrapper here too would double it
-        // (#2900). The AL-visible string is unchanged; it is composed one layer out now.
-        var msg = $"The value must be {comparison} "
-            + $"{boundText}. Value: {FormatValue(value, isInteger)}.";
-
-        var t = System.Type.GetType(
-            "Microsoft.Dynamics.Nav.Types.Exceptions.NavNCLDialogException, Microsoft.Dynamics.Nav.Types");
-        if (t != null)
-        {
-            var ctor = t.GetConstructor(new[] { typeof(string) });
-            if (ctor != null) return (System.Exception)ctor.Invoke(new object[] { msg });
-        }
-        return new System.InvalidOperationException(msg);
-    }
-}
-
-/// <summary>
-/// A Decimal-typed control always renders with the field's decimal places -- default 2,
-/// the same convention #2490 measured against real BC and <see cref="TestPageMinMaxValue.FormatValue"/>
-/// already codifies for the error-message text -- regardless of how many decimal digits the
-/// underlying .NET <c>decimal</c>'s own <c>Scale</c> happens to carry (issues #2634 / #2534).
-///
-/// A record field reaches its stored value through <see cref="Microsoft.Dynamics.Nav.Runtime.NavRecord"/>'s
-/// own Validate/Insert path, which is BC's own precompiled code and out of this runner's
-/// control -- so a Rec-bound Decimal control can format correctly today by construction, if
-/// BC's own write path already normalises the stored scale. A page-GLOBAL <c>Decimal</c>
-/// (<c>Values: array[20] of Decimal;</c>) gets no such round trip at all: a plain assignment
-/// like <c>Values[i] := i * 10;</c> stores a .NET decimal with Scale 0, and
-/// <c>Convert.ToString(10m)</c> answers "10" where real BC's page layer answers "10.00". Both
-/// <see cref="LiveNavTestField.Value"/> and <see cref="PageVariableTestField.Value"/> read
-/// through this helper so neither binding shape can silently regress relative to the other --
-/// exactly the LiveNavTestField/PageVariableTestField pairing pattern <see cref="TestPageBooleanValue"/>
-/// and <see cref="TestPageOptionValue"/> already use.
-///
-/// Only <see cref="NavDecimal"/> needs special handling: <see cref="NavInteger"/> and
-/// <see cref="NavBigInteger"/> already round-trip correctly through
-/// <c>Convert.ToString</c> because an integral CLR type never carries a fractional Scale to
-/// lose in the first place -- matching the "0" (no decimals) half of
-/// <see cref="TestPageMinMaxValue.FormatValue"/>'s own convention without any code needed here.
-///
-/// <para><b>The number of decimals is the CONTROL's, not a constant (#3406).</b> This helper
-/// used to apply <c>"0.00"</c> to every Decimal on every page, so a control declaring
-/// <c>DecimalPlaces = 3 : 3</c> or <c>AutoFormatType = 11</c> read back with two decimals and
-/// nothing said so. The format string is BC's own: <c>NavForm.GetDecimalString</c> runs to
-/// completion inside the runner and publishes its result as the control's
-/// <c>Control&lt;id&gt;_Format</c> source expression, which
-/// <see cref="RunnerPageInstance.TryGetControlFormat"/> reads. See
-/// docs/limitations.md#testpage-decimal-formatting for the measured format strings and what
-/// still is not covered.</para>
-/// </summary>
-internal static class TestPageNumericValue
-{
-    // The two-decimal spelling used when the control has no format of its own — a
-    // record-only field, or a page whose format expression could not be read. Historical:
-    // every existing assertion in the corpus and in Microsoft's own buckets is written
-    // against it, so this arm must not move.
-    private const string NoControlFormat = "0.00";
-
-    /// <param name="controlFormat">
-    /// The .NET format string BC's own <c>GetDecimalString</c> cascade computed for this
-    /// control, or null/empty when there is none to read. See the class remarks: this is
-    /// BC's answer, not the runner's — the runner only applies it.
-    /// </param>
-    internal static string? Format(NavValue? navValue, string? controlFormat = null)
-    {
-        if (navValue is not NavDecimal d) return null;
-
-        var value = Convert.ToDecimal(d.ClientObject, CultureInfo.InvariantCulture);
-        if (string.IsNullOrEmpty(controlFormat))
-            return value.ToString(NoControlFormat, CultureInfo.InvariantCulture);
-
-        // .NET does not reject an unrecognised custom format — it treats the whole string as
-        // a literal and answers it back verbatim, which would put the format string itself
-        // where a number belongs. That is a silent wrong answer of exactly the kind #3406 is
-        // about, so detect it (no digit in the output) and fall back rather than ship it.
-        var formatted = value.ToString(controlFormat, CultureInfo.InvariantCulture);
-        foreach (var c in formatted)
-            if (char.IsDigit(c)) return formatted;
-
-        return value.ToString(NoControlFormat, CultureInfo.InvariantCulture);
-    }
-}
-
-/// <summary>
-/// How a BLANK temporal control renders: the empty string, for Date, Time and DateTime alike.
-///
-/// <para>Issue #2361. The runner fell through to <c>Convert.ToString(ObjectValue)</c>, which
-/// renders the underlying <c>DateTime</c> and so answered <c>01/01/0001 00:00:00</c> where real
-/// BC answers <c>''</c> — the same string for all three types, because all three wrap one
-/// <c>DateTime</c> (see below). Microsoft's own <c>UserCardTest.GenerateWebServiceKeyNoExpires</c>
-/// asserts <c>AssertEquals('')</c> on page 9807's blank <c>WebServiceExpiryDate</c>.</para>
-///
-/// <para><b>The rule is read out of BC, not inferred.</b> Each of BC's three temporal formatters
-/// opens with the identical guard — <c>NavDateFormatter</c>, <c>NavTimeFormatter</c> and
-/// <c>NavDateTimeFormatter.FormatWithFormatNumber</c> in <c>Ncl.dll</c>, all three:</para>
-/// <code>
-///   if (navX.IsZeroOrEmpty) return string.Empty;
-/// </code>
-/// <para>and <c>NavDateTimeValue.IsZeroOrEmpty</c> — the base class of all of <c>NavDate</c>,
-/// <c>NavTime</c> and <c>NavDateTime</c> — is <c>Value == NavDateTimeHelper.DateTimeUndefined</c>,
-/// where <c>DateTimeUndefined</c> is <c>default(DateTime)</c>. So the predicate below is BC's
-/// own, spelled the same way.</para>
-///
-/// <para><b>Why <c>default(DateTime)</c> is never a legitimate value to suppress.</b> BC's
-/// smallest representable Date is <c>DateTimeMinimum</c> = <c>0001-01-02</c>, one day after the
-/// CLR minimum, and its C/SIDE encoding maps 0 to undefined and 1 to that minimum. A real
-/// temporal therefore cannot BE <c>default(DateTime)</c>, so blanking it loses nothing —
-/// which is what lets the populated half of the corpus suite keep passing.</para>
-///
-/// <para>Read through by BOTH <see cref="LiveNavTestField"/> (Rec-bound) and
-/// <see cref="PageVariableTestField"/> (page-global), the same pairing
-/// <see cref="TestPageBooleanValue"/> and <see cref="TestPageOptionValue"/> already use, so
-/// neither binding shape can drift from the other — the corpus suite asserts both.</para>
-///
-/// <para>It must also be what <c>ValueToString</c> answers, for the reason spelled out in
-/// <see cref="TestPageBooleanValue"/>: <c>NavTestField.ALAssertEquals</c> converts the EXPECTED
-/// value through <c>ValueToString</c> and compares it ORDINALLY against the getter, so moving
-/// the getter alone would leave <c>AssertEquals('')</c> failing — which is a distinct test in
-/// the corpus suite and was failing in exactly that way before this fix.</para>
-/// </summary>
-internal static class TestPageBlankTemporalValue
-{
-    /// <summary>The stored NavValue's rendering — <c>""</c> when blank, otherwise null to let
-    /// the caller's existing chain render it.</summary>
-    internal static string? Format(NavValue? navValue)
-        => navValue is NavDateTimeValue t && t.IsZeroOrEmpty ? string.Empty : null;
-
-    /// <summary>The same rule for an already-unwrapped CLR value, as ValueToString sees it —
-    /// <c>ClientObject</c> on all three types hands back the bare <c>DateTime</c>.</summary>
-    internal static string? FormatObject(object? value)
-        => value is DateTime dt && dt == default ? string.Empty : null;
-}
-
-internal static class TestPageBooleanValue
-{
-    /// <summary>
-    /// How a Boolean control RENDERS. Issue #2795: real BC answers "Yes"/"No", measured on all
-    /// eight BC legs of the corpus CI (27.0 through 28.4, run 33967745688 on corpus PR #150 —
-    /// <c>Actual:&lt;Yes&gt;</c> on every failing leg, no other value anywhere in the run) and
-    /// pinned upstream by <c>BooleanFieldControl_ReadsAsYesOrNo</c>. The runner answered
-    /// <c>Convert.ToString(bool)</c>, i.e. "True"/"False".
-    ///
-    /// <para>Read through by BOTH <see cref="LiveNavTestField"/> (a Rec-bound control) and
-    /// <see cref="PageVariableTestField"/> (a page-global one), the same pairing
-    /// <see cref="TestPageNumericValue"/> and <see cref="TestPageOptionValue"/> already use, so
-    /// neither binding shape can drift from the other.</para>
-    ///
-    /// <para>It is also what <c>ValueToString</c> must answer, and that is not a nicety.
-    /// <c>NavTestField.ALAssertEquals</c> — BC's own precompiled method — converts a non-string
-    /// expected value through <c>testField.ValueToString</c> and then compares it ORDINALLY
-    /// against the control's value:</para>
-    /// <code>
-    ///   value = NavValue.CreateNavValueFromObject(NavValueMetadata.DefaultMetadata(testField.FieldType), value);
-    ///   text  = testField.ValueToString(value.ClientObject);
-    ///   if (string.CompareOrdinal(ALValue, text) != 0) throw ...
-    /// </code>
-    /// <para>So changing the getter alone would have broken every
-    /// <c>AssertEquals(&lt;Boolean&gt;)</c> — "Yes" against "True" — which passes today only
-    /// because both halves are wrong in the same way.</para>
-    /// </summary>
-    internal static string? Format(NavValue? navValue)
-        => navValue is NavBoolean b
-            ? (Convert.ToBoolean(b.ClientObject, CultureInfo.InvariantCulture) ? "Yes" : "No")
-            : null;
-
-    /// <summary>The same rendering for an already-unwrapped CLR value, as ValueToString sees it.</summary>
-    internal static string? FormatObject(object? value)
-        => value is bool b ? (b ? "Yes" : "No") : null;
-
-    /// <summary>
-    /// The inverse: the text a TestPage write carries, back to a Boolean.
-    ///
-    /// <para>Accepts "Yes"/"No" ONLY. That is what <see cref="FormatObject"/> now produces, so
-    /// <c>SetValue(&lt;Boolean&gt;)</c> round-trips through it — see the chain in
-    /// <c>NavTestField.ALSetValue</c>, where a non-string value goes out through
-    /// <c>ValueToString</c> and comes back in through this.</para>
-    ///
-    /// <para><b>"True"/"False" is refused, and that is measured, not assumed.</b> An earlier
-    /// version of this fix accepted it, reasoning that it is the spelling AL's own
-    /// <c>Evaluate</c> takes for a Boolean. Corpus PR #163 put the question in front of a
-    /// service tier and all eight BC legs answered identically:</para>
-    /// <code>
-    ///   Validation error for Field: RecTrue,  Message = 'Your entry of 'False' is not an
-    ///   acceptable value for 'Rec True'. (Select Refresh to discard errors)'
-    /// </code>
-    /// <para>So this is not an unsupported surface the runner should refuse as out of scope —
-    /// BC has a defined answer for it, and the runner's job is to give the same one. Hence a
-    /// validation error in BC's own shape rather than a <c>RunnerOutOfScopeException</c>, built
-    /// the same way <see cref="TestPageMinMaxValue.MakeError"/> already builds that shape for a
-    /// MinValue/MaxValue refusal.</para>
-    ///
-    /// <para>One fidelity gap, stated rather than hidden: BC puts the control's declared NAME in
-    /// the <c>Field:</c> slot and its CAPTION in the quoted target ("RecTrue" and "Rec True"
-    /// above). This runner's <c>ITestField.Name</c> answers the caption, so both slots read the
-    /// caption here. A test asserting the message as a substring — as the corpus one does — is
-    /// unaffected; one asserting it verbatim would see the difference.</para>
-    /// </summary>
-    internal static NavValue Resolve(string value, string caption)
-    {
-        if (string.Equals(value, "Yes", StringComparison.OrdinalIgnoreCase)) return NavBoolean.Create(true);
-        if (string.Equals(value, "No", StringComparison.OrdinalIgnoreCase)) return NavBoolean.Create(false);
-
-        throw MakeNotAcceptableError(value, caption);
-    }
-
-    /// <summary>
-    /// BC's own refusal for a value a control will not take:
-    /// <c>Your entry of '{value}' is not an acceptable value for '{caption}'.</c>
-    /// <para>Two layers are deliberately absent here (#2900), because neither is this helper's
-    /// to add: <see cref="TestFieldValidationErrors"/> appends
-    /// <c>" (Select Refresh to discard errors)"</c> when it records the refusal, and BC's own
-    /// <c>NavTestField.CheckError</c> then wraps it in <c>Validation error for Field: {name},
-    /// Message = '…'</c> — including the double space after the comma, which is BC's and not a
-    /// typo. The composed result is the string corpus PR #163 measured on all eight legs.</para>
-    /// </summary>
-    private static System.Exception MakeNotAcceptableError(string value, string caption)
-    {
-        // The BARE message only — see TestPageMinMaxValue.MakeError for why the
-        // "Validation error for Field: ..." wrapper is BC's to add and no longer ours (#2900).
-        var msg = $"Your entry of '{value}' "
-            + $"is not an acceptable value for '{caption}'.";
-
-        var t = System.Type.GetType(
-            "Microsoft.Dynamics.Nav.Types.Exceptions.NavNCLDialogException, Microsoft.Dynamics.Nav.Types");
-        if (t != null)
-        {
-            var ctor = t.GetConstructor(new[] { typeof(string) });
-            if (ctor != null) return (System.Exception)ctor.Invoke(new object[] { msg });
-        }
-        return new System.InvalidOperationException(msg);
-    }
-}
-
-/// <summary>
-/// The Date / DateTime / Time a TestPage control was handed, as a typed NavValue (#3384).
-///
-/// <para>AL's TestPage surface is string-typed for every control, so a temporal value reaches a
-/// control as text by two routes and both end here. A typed AL argument
-/// (<c>SetValue(&lt;Date&gt;)</c>) is rendered to text by BC's own <c>NavTestField.ALSetValue</c>
-/// through <see cref="ITestField.ValueToString"/> before the control ever sees it; text the test
-/// wrote itself (<c>SetValue(Format(D))</c>, <c>SetValue('2026-01-15')</c>) arrives unchanged.</para>
-///
-/// <para>Step two is BC's own client-side evaluator rather than a reimplementation of it, which
-/// is what makes the spellings accepted here the spellings a real service tier accepts.</para>
-/// </summary>
-internal static class TestPageTemporalValue
-{
-    /// <summary>
-    /// Interpret <paramref name="value"/> as <paramref name="type"/>, or answer false and leave
-    /// the caller on its normal <c>NavText</c> path — where BC raises its own refusal, which is
-    /// the message a test asserting a rejected value is written against.
-    /// </summary>
-    internal static bool TryResolve(NavType type, string value, out NavValue? resolved)
-    {
-        resolved = null;
-        if (type is not (NavType.Date or NavType.DateTime or NavType.Time)) return false;
-
-        return TryResolveRoundTrip(type, value, out resolved)
-            || TryEvaluateThroughBc(type, value, out resolved);
-    }
-
-    /// <summary>
-    /// Step one: the spelling this runner's OWN <c>ValueToString</c> produces for a typed
-    /// argument — <c>Convert.ToString(&lt;DateTime&gt;, InvariantCulture)</c>, the general
-    /// date/time pattern <c>MM/dd/yyyy HH:mm:ss</c>. A <c>Time</c> arrives with
-    /// <c>NavTime</c>'s base date attached ("01/02/0001 14:30:00"), which is why the Time arm
-    /// reads <c>TimeOfDay</c> and drops the carrier date.
-    ///
-    /// <para>Handling it here rather than leaving it to BC is not a divergence, and that was
-    /// measured rather than assumed (PR #3394 review): a real BC 28.4.53241.0 tier handed
-    /// <c>SetValue('01/01/2024 00:00:00')</c> as TEXT on a Date control accepts it and stores
-    /// <c>2024-01-01</c> — the same answer this branch gives. What this branch buys is that the
-    /// typed-argument path does not depend on that, since BC's own client never emits this
-    /// spelling for a Date or a Time.</para>
-    ///
-    /// <para><c>TryParseExact</c>, never a lenient parse, and that is the point of the arm:
-    /// a lenient invariant parse would also swallow <c>15.01.28</c> and <c>011528</c> here and
-    /// answer before BC's evaluator ever sees them, silently substituting .NET's reading of a
-    /// user-typed date for the platform's. See <c>TestPageTemporalValueTests</c>.</para>
-    /// </summary>
-    internal static bool TryResolveRoundTrip(NavType type, string value, out NavValue? resolved)
-    {
-        resolved = null;
-        if (type is not (NavType.Date or NavType.DateTime or NavType.Time)) return false;
-
-        if (!DateTime.TryParseExact(value, "MM/dd/yyyy HH:mm:ss", CultureInfo.InvariantCulture,
-                System.Globalization.DateTimeStyles.None, out var roundTrip))
-            return false;
-
-        // Create requires DateTimeKind.Local on all three — the ctors throw
-        // NavNCLDateInvalidException otherwise, and TryParseExact answers Unspecified.
-        resolved = type switch
-        {
-            NavType.Date => NavDate.Create(
-                DateTime.SpecifyKind(roundTrip.Date, DateTimeKind.Local)),
-            NavType.DateTime => NavDateTime.Create(
-                DateTime.SpecifyKind(roundTrip, DateTimeKind.Local)),
-            _ => NavTime.Create(DateTime.SpecifyKind(
-                NavTimeBaseDate.Add(roundTrip.TimeOfDay), DateTimeKind.Local)),
-        };
-        return resolved != null;
-    }
-
-    // NavTime's ClientObject is a DateTime on BC's own base date; Create rejects anything else.
-    private static readonly DateTime NavTimeBaseDate = new DateTime(1, 1, 2);
-
-    private static bool _lookupDone;
-    private static System.Reflection.MethodInfo? _getEvaluator;
-    private static System.Reflection.MethodInfo? _evaluate;
-    private static object? _trapError;
-    private static Type? _navNclType;
-
-    private static bool TryEvaluateThroughBc(NavType type, string value, out NavValue? resolved)
-    {
-        resolved = null;
-        EnsureEvaluatorBound();
-
-        var member = type switch
-        {
-            NavType.Date => "NavDate",
-            NavType.DateTime => "NavDateTime",
-            _ => "NavTime",
-        };
-        // Both pre-invoke steps THROW rather than decline (#3560). Declining sends the caller
-        // to its NavText path, where BC raises its date-format refusal - so an AL author reads
-        // BC rejecting the spelling they typed when what happened is that the runner could not
-        // map this control's type onto this build's NavNclType at all.
-        if (ForcedEnumMemberMissing.Value || !Enum.IsDefined(_navNclType!, member))
-            throw new AlRunner.Infrastructure.BcShapeGapException(
-                "TestPage SetValue on a Date/DateTime/Time control",
-                "NavNclType." + member,
-                $"this build's NavNclType does not define {member}, so the runner cannot name "
-                + "the type whose evaluator it needs to read this control's value. A "
-                + "runner/BC-version mismatch, not a rejected value.");
-
-        var evaluator = ForcedNullEvaluator.Value
-            ? null
-            : _getEvaluator!.Invoke(null, new[] { Enum.Parse(_navNclType!, member) });
-        if (evaluator == null)
-            throw new AlRunner.Infrastructure.BcShapeGapException(
-                "TestPage SetValue on a Date/DateTime/Time control",
-                "NavValueEvaluator.GetEvaluator",
-                $"GetEvaluator answered null for NavNclType.{member}, so this build declares "
-                + "the type but hands the runner no evaluator for it and the runner has no way "
-                + "to ask BC how it reads this value. A runner/BC-version mismatch, not a "
-                + "rejected value.");
-
-        // DataError.TrapError, not ThrowError: a spelling BC cannot read has to come back as
-        // "no" so the caller keeps its NavText path, where BC raises the refusal AL is written
-        // against. Throwing from here would replace that message with this one.
-        var args = new object?[] { BcRuntime.SkeletonSession, _trapError, null, null, value, 0 };
-        if (!InvokeEvaluate(() => (bool)_evaluate!.Invoke(evaluator, args)!)) return false;
-        resolved = args[2] as NavValue;
-        return resolved != null;
-    }
-
-    /// <summary>
-    /// Run BC's evaluator and answer what it answered - but only for an answer BC gave.
-    ///
-    /// <para>Under <c>DataError.TrapError</c> a refusal is not an exception: BC returns false and
-    /// traps its own <c>NavNCLEvaluateException</c> (28.1 Ncl, <c>NavValueEvaluator`1.Evaluate</c>:
-    /// <c>catch when (obj is NavNCLEvaluateException &amp;&amp; dataError == 0)</c>; TrapError is
-    /// 0). So anything else reaching here is the evaluator NOT COMPLETING, and answering false for
-    /// it shows the AL author BC's date-format refusal for a value BC would have taken (#3444).</para>
-    ///
-    /// <para><c>BcShapeGapException</c> because an <c>asserterror</c> can absorb a
-    /// <c>RunnerOutOfScopeException</c> and would then pass on a runner fault (#3062, #3428).</para>
-    /// </summary>
-    internal static bool InvokeEvaluate(Func<bool> invoke)
-    {
-        try { return invoke(); }
-        catch (System.Reflection.TargetInvocationException ex)
-            when (ex.InnerException is NavNCLEvaluateException)
-        {
-            // BC ran and said no, in the one shape that can still arrive as a throw.
-            return false;
-        }
-        catch (System.Reflection.TargetInvocationException ex)
-        {
-            var inner = ex.InnerException ?? ex;
-            throw new AlRunner.Infrastructure.BcShapeGapException(
-                "TestPage SetValue on a Date/DateTime/Time control",
-                "NavValueEvaluator.Evaluate",
-                $"BC's own value evaluator did not complete ({inner.GetType().Name}: "
-                + $"{inner.Message}), so the runner has no answer from BC about this spelling. "
-                + "This is a runner fault on the evaluate path, not a value BC rejected.");
-        }
-        catch (Exception ex) when (ex is System.ArgumentException
-                                or System.Reflection.TargetParameterCountException
-                                or System.Reflection.TargetException)
-        {
-            // Unwrapped, so it came from Invoke itself rather than from BC: the bound Evaluate
-            // does not take the arguments this call site passes. The two reflection types are
-            // not ArgumentExceptions — both derive from ApplicationException — so before #3462
-            // they propagated raw past every arm here and an asserterror absorbed them.
-            throw new AlRunner.Infrastructure.BcShapeGapException(
-                "TestPage SetValue on a Date/DateTime/Time control",
-                "NavValueEvaluator.Evaluate",
-                $"the bound Evaluate refused this call's arguments ({ex.GetType().Name}: "
-                + $"{ex.Message}), so this BC build's evaluator is not the shape the runner "
-                + "reflects against. A runner/BC-version mismatch, not a rejected value.");
-        }
-    }
-
-    // Bound by reflection because NavValueEvaluator and NavNclType are internal to Ncl.dll. The
-    // shape was read off the 28.1 and 28.4 decompiles; if a BC build moves it, the binding must
-    // FAIL LOUDLY rather than quietly, because failing quietly is invisible: the typed-argument
-    // arms keep working through the round-trip branch above and only the text spellings revert
-    // to the pre-#3384 refusal. That is a silent downgrade of the kind loud-failures.md exists
-    // to prevent, so it gets one line on stderr, once — not a throw, because declining still
-    // leaves BC's own refusal as the observable outcome rather than a wrong value.
-    private static bool TryBindEvaluator()
-    {
-        // Latched AFTER the work, under a gate: setting _lookupDone first let a second thread
-        // read "already bound" while _evaluate was still null, and EnsureEvaluatorBound then
-        // refused a BC build it had bound perfectly well, naming no reason (#3444, #3187's
-        // latch-before-work shape).
-        lock (BindGate)
-        {
-            if (_lookupDone) return _evaluate != null;
-            try { return BindEvaluatorCore(); }
-            finally { _lookupDone = true; }
-        }
-    }
-
-    private static readonly object BindGate = new();
-
-    private static bool BindEvaluatorCore()
-    {
-        string? why = null;
-        try
-        {
-            var ncl = AppDomain.CurrentDomain.GetAssemblies()
-                .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Ncl");
-            var evaluatorType = ncl?.GetType("Microsoft.Dynamics.Nav.Runtime.NavValueEvaluator");
-            _navNclType = ncl?.GetType("Microsoft.Dynamics.Nav.Runtime.NavNclType");
-            var dataErrorType = typeof(Microsoft.Dynamics.Nav.Types.DataError);
-
-            const System.Reflection.BindingFlags Any =
-                System.Reflection.BindingFlags.Public | System.Reflection.BindingFlags.NonPublic
-                | System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.Instance;
-
-            if (ncl == null) why = "Microsoft.Dynamics.Nav.Ncl is not loaded";
-            else if (evaluatorType == null) why = "NavValueEvaluator not found";
-            else if (_navNclType is not { IsEnum: true }) why = "NavNclType not found, or not an enum";
-            else if (!Enum.IsDefined(dataErrorType, "TrapError")) why = "DataError.TrapError not found";
-            else
-            {
-                _getEvaluator = evaluatorType.GetMethod(
-                    "GetEvaluator", Any, null, new[] { _navNclType }, null);
-                // BcShape.FindMethod, not GetMethod(name, flags): the latter throws a bare
-                // AmbiguousMatchException naming no member the moment BC ships a second
-                // Evaluate, and NavMethodScope_AssertError rethrows only BcShapeGapException —
-                // so under an AL asserterror that one would be ABSORBED and the asserterror
-                // would pass (#3069). The signature cannot be pinned here because four of the
-                // six parameter types are internal to Ncl, so this resolves by name and refuses
-                // a second declaration by name, which is the outcome the ambiguity guard wants.
-                var evaluate = AlRunner.Infrastructure.BcShape.FindMethod(
-                    evaluatorType, "Evaluate", Any,
-                    "TestPage SetValue on a Date/DateTime/Time control",
-                    "NavValueEvaluator.Evaluate",
-                    "the runner reads a control's typed date through BC's own value evaluator");
-
-                if (_getEvaluator == null) why = "NavValueEvaluator.GetEvaluator(NavNclType) not found";
-                else if (evaluate == null) why = "NavValueEvaluator.Evaluate not found";
-                else if (evaluate.GetParameters().Length != 6)
-                    why = $"NavValueEvaluator.Evaluate takes {evaluate.GetParameters().Length} "
-                        + "parameters, expected 6";
-                else
-                {
-                    _trapError = Enum.Parse(dataErrorType, "TrapError");
-                    _evaluate = evaluate;
-                    return true;
-                }
-            }
-        }
-        catch (Exception ex)
-        {
-            why = $"{ex.GetType().Name}: {ex.Message}";
-        }
-
-        _evaluate = null;
-        _bindFailure = why;
-        return false;
-    }
-
-    private static string? _bindFailure;
-
-    // TEST SEAM (#3462). The bind result latches once per process, so a test cannot provoke a
-    // real bind failure without destroying the binding for every other test in the run. This
-    // makes EnsureEvaluatorBound refuse as though the bind had failed, without touching the
-    // latch. AsyncLocal, not a plain static: xunit runs test classes in parallel and the
-    // temporal suites drive this path concurrently.
-    private static readonly System.Threading.AsyncLocal<string?> ForcedBindFailure = new();
-
-    internal static IDisposable ForceBindFailure(string reason)
-    {
-        var previous = ForcedBindFailure.Value;
-        ForcedBindFailure.Value = reason;
-        return new ForcedBindFailureScope(previous);
-    }
-
-    // TEST SEAMS (#3560) for the two pre-invoke declines below EnsureEvaluatorBound. Neither is
-    // reachable on a BC build the runner supports - both the enum member and its evaluator
-    // resolve - so a refusal nobody can provoke would ship untested. AsyncLocal for the same
-    // reason ForcedBindFailure is: the temporal suites run in parallel.
-    private static readonly System.Threading.AsyncLocal<bool> ForcedEnumMemberMissing = new();
-    private static readonly System.Threading.AsyncLocal<bool> ForcedNullEvaluator = new();
-
-    internal static IDisposable ForceEnumMemberMissing() => Force(ForcedEnumMemberMissing);
-
-    internal static IDisposable ForceNullEvaluator() => Force(ForcedNullEvaluator);
-
-    private static IDisposable Force(System.Threading.AsyncLocal<bool> flag)
-    {
-        var previous = flag.Value;
-        flag.Value = true;
-        return new ForcedFlagScope(flag, previous);
-    }
-
-    private sealed class ForcedFlagScope : IDisposable
-    {
-        private readonly System.Threading.AsyncLocal<bool> _flag;
-        private readonly bool _previous;
-        internal ForcedFlagScope(System.Threading.AsyncLocal<bool> flag, bool previous)
-        {
-            _flag = flag;
-            _previous = previous;
-        }
-        public void Dispose() => _flag.Value = _previous;
-    }
-
-    private sealed class ForcedBindFailureScope : IDisposable
-    {
-        private readonly string? _previous;
-        internal ForcedBindFailureScope(string? previous) => _previous = previous;
-        public void Dispose() => ForcedBindFailure.Value = _previous;
-    }
-
-    /// <summary>
-    /// Bind BC's evaluator, or THROW — a failure to BIND means the runner cannot ask THIS BC
-    /// build how it reads a date, which is a shape gap and not a decline. Declining would be
-    /// invisible: the typed-argument path keeps working through <see cref="TryResolveRoundTrip"/>
-    /// and only text spellings quietly revert to the pre-#3384 refusal.
-    /// </summary>
-    internal static void EnsureEvaluatorBound()
-    {
-        var forced = ForcedBindFailure.Value;
-        if (forced == null && TryBindEvaluator()) return;
-
-        // BcShapeGapException, not RunnerOutOfScopeException: an AL asserterror absorbs the
-        // latter (NavMethodScope_AssertError rethrows only this type), so `asserterror
-        // SetValue(<temporal>)` passed green on a runner that could not ask BC anything (#3462).
-        throw new AlRunner.Infrastructure.BcShapeGapException(
-            "TestPage SetValue on a Date/DateTime/Time control",
-            "NavValueEvaluator.Evaluate",
-            "could not bind BC's own NavValueEvaluator ("
-            + (forced ?? _bindFailure ?? "reason not recorded") + "), so the runner cannot ask this BC "
-            + "build how it reads a date, time or datetime a test typed as text. This is a "
-            + "runner/BC-version mismatch, not a rejected value.");
-    }
-}
-
-internal sealed class LiveNavTestField : ITestField
-{
-    private readonly NavRecord _record;
-    private readonly int _fieldNo;
-    // The page behind the control, when there is one. A Rec-bound control still has an
-    // OnLookup trigger on the page, and that trigger is the only thing Lookup() can run.
-    private readonly RunnerPageInstance? _page;
-    private readonly int _controlId;
-
-    // Told when this field writes, so the page can persist the row at the moment BC would.
-    // The field itself owns no page state and must not: a part's fields belong to the part's
-    // page, not to the card the test is holding.
-    private readonly Action? _onEdited;
-
-    // Told BEFORE this field writes, so the page can turn its implicit new-row line into a
-    // real started row while the write's own OnValidate can still see it — see
-    // LiveNavTestPage.PromoteNewRowLineForWrite (#2923). Separate from _onEdited because the
-    // order is the whole point: _onEdited runs after the validate and is far too late to give
-    // the row its keys.
-    private readonly Action? _onBeforeEdit;
-
-    public LiveNavTestField(NavRecord record, int fieldNo)
-        : this(record, fieldNo, page: null, controlId: 0, onEdited: null, onBeforeEdit: null,
-               pageValidationErrors: null) { }
-
-    public LiveNavTestField(NavRecord record, int fieldNo, RunnerPageInstance? page, int controlId,
-        Action? onEdited, Action? onBeforeEdit, TestPageValidationErrors? pageValidationErrors)
-    {
-        _record = record;
-        _fieldNo = fieldNo;
-        _page = page;
-        _controlId = controlId;
-        _onEdited = onEdited;
-        _onBeforeEdit = onBeforeEdit;
-        _validationErrors = new TestFieldValidationErrors(pageValidationErrors);
-    }
-
-    // The refusals this control has recorded, read back by ValidationErrorCount /
-    // GetValidationError below. See TestFieldValidationErrors for BC's own contract: the
-    // ITestField setter RECORDS a refusal, it does not throw it — NavTestField.CheckError
-    // (BC's own precompiled code, wrapping every SetValue) is what raises it afterwards, and
-    // it can only do that if the ledger survives the write (#2900).
-    //
-    // Constructed with the owning PAGE's ledger (#3009) so a refusal lands in both: BC reads
-    // the field's count AND the page's around every write, and AL can read either afterwards.
-    // Null for the record-only ctor above, which has no page to report to.
-    private readonly TestFieldValidationErrors _validationErrors;
-
-    public string Value
-    {
-        // An option field answers with its MEMBER NAME, not the ordinal it stores. Returning the
-        // ordinal made every comparison against a member name fail while looking like a data
-        // problem ("expected <Mid>, got <0>") rather than a missing option table.
-        get => (CurrentOption() is { } option
-                   ? TestPageOptionValue.Display(option, OptionCaptions())
-                   : null)
-               // #3406: the control's OWN decimal format, computed by BC's GetDecimalString
-               // cascade and read off the page, not a hardcoded two decimals. Null _page is
-               // the record-only field, which has no control and keeps the old spelling.
-               ?? TestPageNumericValue.Format(_record.GetFieldValue(_fieldNo) as NavValue,
-                                              _page?.TryGetControlFormat(_controlId))
-               // #2795: "Yes"/"No", not Convert.ToString's "True"/"False".
-               ?? TestPageBooleanValue.Format(_record.GetFieldValue(_fieldNo) as NavValue)
-               // #2361: a blank Date/Time/DateTime is '', not the rendered CLR minimum.
-               ?? TestPageBlankTemporalValue.Format(_record.GetFieldValue(_fieldNo) as NavValue)
-               ?? Convert.ToString(ObjectValue, CultureInfo.InvariantCulture)
-               ?? string.Empty;
-        // appendRefreshSuffix: true — a Rec-bound control stages a row edit, and real BC's
-        // client decorates its recorded validation error with the offer to discard it. Measured
-        // on corpus run 34002487601; see TestFieldValidationErrors' header.
-        set => _validationErrors.RunRecordingRefusal(() => Write(value), appendRefreshSuffix: true);
-    }
-
-    private void Write(string value)
-    {
-        // FIRST, before anything reads or writes the record: if the page is parked on its
-        // implicit new-row line, this write is what turns that line into a row the flush path
-        // will persist, and BC settles the row's key BEFORE the typed value is validated onto
-        // it. Doing it after (which is all MarkEdited below could do) handed the field's own
-        // OnValidate a row with no key: issue #2923, 35 Tests-SMB failures on Sales Line and
-        // Purchase Line. A no-op on every page not sitting on that line, which is all of them
-        // once a real row is current.
-        //
-        // Inside Write, so it is inside the RunRecordingRefusal the setter wraps this in
-        // (#3007): starting the row is part of this control write, so a refusal raised while
-        // starting it is recorded and re-raised by BC's NavTestField.CheckError exactly as a
-        // refusal of the value itself would be.
-        _onBeforeEdit?.Invoke();
-
-        // #3640: everything from here on can mutate Rec — the field's own OnValidate, the
-        // control's, and any pageextension modify() trigger around them — and real BC discards
-        // those mutations when the write raises. Snapshot AFTER _onBeforeEdit, because the
-        // new-row promotion above writes the row's KEY, which is page state settled before the
-        // value is validated rather than a trigger's mutation of it; the tier measured the
-        // trigger half and says nothing about unwinding the promotion.
-        //
-        // Taken and put back INLINE rather than by wrapping the rest of this method in a
-        // lambda: TestPageNewRowLinePromotionTests reads this method's IL to pin that
-        // _onBeforeEdit precedes ALValidateAsync and _onEdited follows it (#2923), and moving
-        // any of those three into a compiler-generated closure hides the ordering from the
-        // one test that guards it. See TestPageWriteBuffer.
-        var restore = TestPageWriteBuffer.Snapshot(_record);
-        try
-        {
-        // Issue #1870 — the Rec-bound half of #1837 that #1869 (the page-variable half)
-        // left open. FieldType (sourced from the source table field's own declared type,
-        // see TryGetMetaFieldType) answers Boolean for a `field(Flag; Rec.Flag)` control
-        // over a `Boolean` table field; falling through to ALCompiler.ToNavValue(value)
-        // there always produced a NavText, which NavTestField.ALSetValue's own Boolean
-        // ALValidateAsync then rejected with "The value 'True' can't be evaluated into
-        // type Boolean" — the same shape of bug TestPageBooleanValue already fixed for
-        // PageVariableTestField.
-        //
-        // #3384 — Date/DateTime/Time needed the same kind of arm. Without one this fell through
-        // to ToNavValue for every temporal control, so the typed value BC's ALSetValue had just
-        // rendered to text was validated as a NavText and refused, and text the test wrote
-        // itself was refused the same way — including the ISO spelling BC's own message
-        // recommends. See TestPageTemporalValue.
-        var navValue = CurrentOption() is { } option
-            ? TestPageOptionValue.Resolve(option, value, OptionCaptions(),
-                $"TestPage SetValue (field {_fieldNo})")
-            : FieldType == NavType.Boolean
-                ? TestPageBooleanValue.Resolve(value, Caption)
-                : TestPageTemporalValue.TryResolve(FieldType, value, out var temporal)
-                    ? temporal!
-                    : ALCompiler.ToNavValue(value);
-
-        // MinValue/MaxValue (#2495): measured against real BC (28.1/28.4), a bounded field's
-        // MinValue/MaxValue is enforced on a TestPage control WRITE, but NOT on Rec.Validate
-        // or a plain field assignment — so this check belongs here, at the page-write layer,
-        // and must not move into ALValidateAsync below (that is also what Rec.Validate calls,
-        // and pulling the check in there would enforce it on Validate too, which real BC does
-        // not). See TestPageMinMaxValue.Check's own doc comment for the exact message shape.
-        TestPageMinMaxValue.Check(_record.MetaTable, _fieldNo, FieldType, value, Caption);
-
-        // Setting a field on a page is a VALIDATE, not an assignment. That is what fills in
-        // the caption when a user picks an id, and what lets a field refuse a value outright.
-        // A raw SetFieldValue stored what the test wrote — so the field itself read back
-        // correctly and every field DERIVED from it stayed empty, which made the test fail
-        // pointing at the derived field, the one place the defect was not.
-        //
-        // Issue #2705 — real BC (measured on a 28.4 container) runs the bound field's
-        // OnValidate with CurrFieldNo equal to that field's number for the duration of a
-        // page-driven write (own-table AND tableextension fields alike), while a
-        // Rec.Validate from AL code leaves it at 0. NavRecord.CurrFieldNo
-        // (Microsoft.Dynamics.Nav.Ncl.dll, decompiled) is a plain public get/set property
-        // that nothing in Ncl itself ever assigns — real BC's compiled client/page glue must
-        // set it around a UI-originated validate, which is exactly what a TestPage SetValue
-        // is standing in for here. Restoring the PREVIOUS value (not unconditionally 0)
-        // keeps a nested SetValue-from-OnValidate honest, and the try/finally matches what
-        // arm E of the corpus test measures: OnModify after Close() sees CurrFieldNo = 0
-        // again, so the assignment must not outlive this one validate call.
-        var previousCurrFieldNo = _record.CurrFieldNo;
-        _record.CurrFieldNo = _fieldNo;
-        // #3543: BC brackets a page-driven field validate in a transaction of its own —
-        // NavRecord.ValidateFieldsAsync opens Session.BeginTransaction() per field and closes
-        // it with Session.EndTransaction(commit) in a finally, and NavForm.ModifyAsync does
-        // the same around the page's own Modify. So an OnValidate that writes is legal even
-        // when the CALLER holds no transaction, which under TransactionModel::None is the
-        // whole test body (#3480). Without this bracket the runner refused a write BC allows.
-        //
-        // Observably equivalent: the counter grants a write nothing outside a None scope,
-        // where ThrowIfNoTransactionForWrite returns before it is read at all, and it changes
-        // no commit point — which is a separate question, tracked by NoteTransactionEnd and
-        // deliberately left alone here (#2413 measured that conflating the two is wrong).
-        AlRunner.Patches.ALDatabasePatches.EnterRunTransaction();
-        try
-        {
-            try
-            {
-                _record.ALValidateAsync(_fieldNo, navValue, null).GetAwaiter().GetResult();
-            }
-            finally
-            {
-                _record.CurrFieldNo = previousCurrFieldNo;
-            }
-
-            // Then the control's own OnValidate, which is a second and independent trigger: the
-            // table field's runs first, the page's after it.
-            if (_page != null && _controlId != 0) _page.RaiseOnValidate(_controlId);
-        }
-        finally
-        {
-            AlRunner.Patches.ALDatabasePatches.ExitRunTransaction();
-        }
-
-        _onEdited?.Invoke();
-        }
-        catch
-        {
-            restore?.Invoke();
-            throw;
-        }
-    }
-
-    // The stored NavValue, not the unwrapped ClientObject — the option metadata rides on the
-    // NavOption itself, and unwrapping it to an int is what loses the member table.
-    private NavOption? CurrentOption() => _record.GetFieldValue(_fieldNo) as NavOption;
-
-    // Record-only mode has no control to carry an OptionCaption, so members are all there is.
-    // CurrentOption() is passed through so an Enum-typed field can fall back to the enum's
-    // own captions when the control declares no OptionCaption — see TryGetOptionCaptions.
-    private string[]? OptionCaptions()
-        => _page != null && _controlId != 0 ? _page.TryGetOptionCaptions(_controlId, CurrentOption()) : null;
-
-    public string Name => Caption;
-
-    // TestPage field Caption() (#1777). BC's own precedence, control-declared wins over the
-    // source field's Caption, which wins over the field's bare name:
-    //   1. the control's own Caption (field(Foo; Rec.Foo) { Caption = '…'; }) — page metadata
-    //      that only exists when this field is bound to a live control, not a bare NavRecord.
-    //   2. the source table field's declared Caption (field(2; Foo; Text[30]) { Caption = '…'; })
-    //      — read straight from the parse-time metadata, bypassing NCLMetaField.FieldCaption
-    //      (JmpHooked to always answer the field NAME; see TryGetParsedFieldCaption).
-    //   3. the field's technical name, BC's own fallback when neither is declared.
-    public string Caption
-        => (_page != null && _controlId != 0 ? _page.TryGetControlCaption(_controlId) : null)
-           ?? TryGetMetaFieldCaption()
-           ?? TryGetMetaFieldName()
-           ?? $"Field {_fieldNo}";
-    public NavType FieldType => TryGetMetaFieldType() ?? NavType.Text;
-    // BC's own NavTestField.CheckError reads all three around every control write, and
-    // NavTestField.ALValidationErrorCount / ALGetValidationError hand the first and fourth
-    // straight to AL. Hardcoded 0/"" made `ValidationErrorCount()` answer 0 after a refusal
-    // real BC counts as 1, and made a refusal escape the setter raw instead of being wrapped
-    // by BC in "Validation error for Field: ..." (#2900). See TestFieldValidationErrors.
-    public int ValidationErrorCount => _validationErrors.Count;
-    public long LastUsedValidationErrorId => _validationErrors.LastUsedId;
-    public long MaxValidationErrorId => _validationErrors.MaxId;
-    public object? ObjectValue => LiveNavTestPage.Unwrap(_record.GetFieldValue(_fieldNo));
-    public int OptionCount => CurrentOption() is { } option ? TestPageOptionValue.Count(option) : 0;
-
-    // The control's declared state, not a constant. `Editable = false` / `Editable = SomeVar`
-    // is how a page protects rows it does not own, so answering true unconditionally made
-    // every test of that protection pass no matter what the page said. Falls back to true
-    // only when there is no page object to ask — the record-only mode, which has no control
-    // metadata at all and never claimed to model these.
-    public bool Enabled  => _page?.ControlEnabled(_controlId) ?? true;
-    public bool Editable => _page?.ControlEditable(_controlId) ?? true;
-    public bool Visible  => _page?.ControlVisible(_controlId) ?? true;
-    public bool HideValue => false;
-    public bool ShowMandatory => false;
-
-    public string GetValidationError(int index) => _validationErrors.Get(index);
-    public void Activate() { }
-
-    /// <summary>
-    /// Run the control's OnLookup trigger — the AL a user's F4 would run. The base mock does
-    /// nothing, which let a test invoke a lookup, observe no change, and compare two empty
-    /// strings successfully.
-    /// </summary>
-    public void Lookup()
-    {
-        if (_page == null)
-            throw TestPageShapeGap.Lookup(
-                $"TestPage lookup on field {_fieldNo}",
-                "no AL page object was built for this page, so its OnLookup trigger cannot be "
-                + "reached");
-
-        // BC's contract: the trigger writes the selection back and returns true; a false
-        // return means the user cancelled and the field keeps its value.
-        // The record and field number go with the call so RaiseOnLookup can fall back to the
-        // SOURCE TABLE FIELD's own OnLookup when the control declares none (#2549). A table
-        // trigger writes into Rec rather than handing a value back, which is why the null it
-        // returns for that path needs no special handling here: the Value getter reads the
-        // record, where the trigger already wrote.
-        var picked = _page.RaiseOnLookup(_controlId, NavText.Create(Value), _record, _fieldNo);
-        if (picked != null) Value = picked.ToString();
-    }
-
-    public void Lookup(NavDataSet dataSet) => Lookup();
-
-    /// <summary>
-    /// Run the control's OnAssistEdit trigger — see RunnerPageInstance.RaiseOnAssistEdit,
-    /// including why a control with no such trigger stays silent here rather than refusing.
-    /// Was a literal no-op (#2362, #3642), so a page's declared OnAssistEdit never ran and
-    /// nothing said so: the UI handlers the test bound for what that trigger raises went
-    /// unexecuted and the test failed at teardown, pointing at the handlers rather than here.
-    /// </summary>
-    public void AssistEdit()
-    {
-        if (_page == null)
-            throw TestPageShapeGap.AssistEdit(
-                $"TestPage assist-edit on field {_fieldNo}",
-                "no AL page object was built for this page, so its OnAssistEdit trigger cannot "
-                + "be reached");
-
-        _page.RaiseOnAssistEdit(_controlId);
-    }
-
-    /// <summary>
-    /// Run the control's OnDrillDown trigger — see RunnerPageInstance.RaiseOnDrillDown for the
-    /// full contract, including the fixed error real BC raises when no trigger is declared.
-    /// Left #57's literal no-op (`public void Drilldown() { }`), which let a test call
-    /// DrillDown(), observe nothing happened, and pass anyway — the trigger's effect (or its
-    /// documented absence-error) never ran, and the test only tripped one step later on a
-    /// missing side effect that pointed at the wrong place.
-    /// </summary>
-    public void Drilldown()
-    {
-        if (_page == null)
-            throw TestPageShapeGap.DrillDown(
-                $"TestPage drilldown on field {_fieldNo}",
-                "no AL page object was built for this page, so its OnDrillDown trigger cannot "
-                + "be reached");
-
-        _page.RaiseOnDrillDown(_controlId);
-    }
-
-    public void Invoke() { }
-
-    // An Option/Enum-bound control renders an ordinal as the text the control SHOWS, the same
-    // spelling the Value getter answers with — issue #2367. BC's own ALAssertEquals/ALSetValue
-    // strip an AL option value down to a bare ordinal before calling this (see
-    // TestPageOptionValue.DisplayOrdinal for the exact chain), so leaving it as
-    // Convert.ToString made AssertEquals compare the ordinal '2' against the control's
-    // 'Pending Approval' and report a mismatch for the value the record actually held.
-    public string ValueToString(object? value)
-        => TestPageOptionValue.DisplayOrdinal(CurrentOption(), value, OptionCaptions())
-           // #2795: BC's ALAssertEquals converts the EXPECTED value through here and compares it
-           // ordinally against the control's Value, so this has to answer with the same word the
-           // getter above does or AssertEquals(<Boolean>) can never match.
-           ?? TestPageBooleanValue.FormatObject(value)
-           // #2361: same reason, for a blank temporal — AssertEquals('') is its own corpus test.
-           ?? TestPageBlankTemporalValue.FormatObject(value)
-           ?? Convert.ToString(value, CultureInfo.InvariantCulture) ?? string.Empty;
-
-    // AL that walks an option set (building a picker, asserting the members a field offers) got
-    // an empty string for every index, which reads as "this option has blank members" rather
-    // than as an unimplemented accessor.
-    public string GetOption(int index)
-        => CurrentOption() is { } option
-            ? TestPageOptionValue.MemberAt(option, index, OptionCaptions())
-            : string.Empty;
-
-    private string? TryGetMetaFieldName()
-    {
-        return _record.MetaTable.TryGetFieldByNo(_fieldNo, out var field) ? field.FieldName : null;
-    }
-
-    // The source field's own declared Caption — see RecordPatches.TryGetParsedFieldCaption
-    // for why this cannot go through NCLMetaField.FieldCaption.
-    private string? TryGetMetaFieldCaption()
-    {
-        var tableId = _record.MetaTable.TableId;
-        return tableId != 0 ? RecordPatches.TryGetParsedFieldCaption(tableId, _fieldNo) : null;
-    }
-
-    private NavType? TryGetMetaFieldType()
-    {
-        return _record.MetaTable.TryGetFieldByNo(_fieldNo, out var field) ? field.FieldNavType : null;
-    }
-}
-
-/// <summary>
-/// A TestPage field over a control bound to a PAGE VARIABLE rather than to a source-table
-/// field. Reads and writes go through the page's own source expression — BC's binding, not
-/// a runner-side copy — so the value lives on the page instance exactly where the AL
-/// declared it, and a second page instance starts with its own.
-///
-/// Writing also runs the control's OnValidate trigger, because that is what setting a
-/// value on a page does; a setter that skipped it would let a test observe the value it
-/// just wrote while none of the page's AL had run.
-/// </summary>
-internal sealed class PageVariableTestField : ITestField
-{
-    private readonly RunnerPageInstance _page;
-    private readonly object _expression;
-    private readonly int _controlId;
-
-    public PageVariableTestField(RunnerPageInstance page, object expression, int controlId,
-        TestPageValidationErrors? pageValidationErrors)
-    {
-        _page = page;
-        _expression = expression;
-        _controlId = controlId;
-        _validationErrors = new TestFieldValidationErrors(pageValidationErrors);
-    }
-
-    // The Rec-bound sibling's ledger, for the same reason and read the same way — see
-    // LiveNavTestField._validationErrors and TestFieldValidationErrors. A page-global control
-    // refuses a write through its OnValidate exactly as a Rec-bound one does, so leaving this
-    // half hardcoded would have made the same AL assertion answer differently depending only
-    // on how the control happens to be bound.
-    //
-    // Constructed with the owning PAGE's ledger (#3009), same as the Rec-bound sibling — the
-    // page counts a refusal whichever way the control that raised it is bound.
-    private readonly TestFieldValidationErrors _validationErrors;
-
-    public string Value
-    {
-        // An Option/Enum-bound control answers with its CAPTION, not the ordinal it stores —
-        // the read-side complement of #1928 (issue #2055). LiveNavTestField.Value already does
-        // this for a Rec-bound control; this class never got it, so `Format(Field.Value())` on
-        // a page-variable enum control returned "1" instead of "OR" while the write direction
-        // (SetValue, below) already resolved captions correctly.
-        get => (CurrentOption() is { } option
-                   ? TestPageOptionValue.Display(option, _page.TryGetOptionCaptions(_controlId, option))
-                   : null)
-               // #3406: the page-global half of the decimal-format rule — see the Rec-bound
-               // sibling. A page-variable control carries the same Control<id>_Format
-               // expression, so the two binding shapes must not answer differently.
-               ?? TestPageNumericValue.Format(RunnerPageInstance.GetValue(_expression),
-                                              _page.TryGetControlFormat(_controlId))
-               // #2795: the page-global half of the same rule — see TestPageBooleanValue.Format.
-               ?? TestPageBooleanValue.Format(RunnerPageInstance.GetValue(_expression))
-               // #2361: the page-global half of the blank-temporal rule — see the Rec-bound
-               // sibling. Base Application page 9807 binds WebServiceExpiryDate this way.
-               ?? TestPageBlankTemporalValue.Format(RunnerPageInstance.GetValue(_expression))
-               ?? Convert.ToString(ObjectValue, CultureInfo.InvariantCulture)
-               ?? string.Empty;
-        // appendRefreshSuffix: false — a page-global control stages no row edit, so there is
-        // nothing for "Refresh to discard" to discard. Microsoft's Tests-SINGLESERVER
-        // Codeunit134614 asserts the bare text with exact equality for exactly this binding
-        // shape (verified mechanically to be page-variable-bound, not Rec-bound). This is the
-        // half no service-tier run has confirmed yet — corpus PR #184 asks it.
-        //
-        // #3640: the Rec-bound sibling's restore-on-refusal applies here too. A page-variable
-        // control's OnValidate is ordinary AL and can write Rec exactly as a Rec-bound one's
-        // can, and a page-driven write is a page-driven write whichever way the CONTROL that
-        // started it happens to be bound — so leaving this half out would make the same AL
-        // observable depend on a binding detail the tier's claim does not mention. Only Rec is
-        // restored: what a failed write leaves in a page GLOBAL is a separate claim no service
-        // tier has measured, and inventing an answer for it is what
-        // ask-the-corpus-before-claiming-bc-behavior.md forbids.
-        set => _validationErrors.RunRecordingRefusal(
-            () => TestPageWriteBuffer.RunRestoringOnRefusal(_page.Record, () =>
-            {
-                RunnerPageInstance.SetValue(_expression, ToBoundValue(value));
-                _page.RaiseOnValidate(_controlId);
-            }), appendRefreshSuffix: false);
-    }
-
-    public object? ObjectValue => LiveNavTestPage.Unwrap(RunnerPageInstance.GetValue(_expression));
-
-    // The stored NavValue, not the unwrapped ClientObject — see LiveNavTestField.CurrentOption
-    // for why: the option metadata (and, for an Enum, whether it IS one — see
-    // TestPageOptionValue.EnumCaptions) rides on the NavOption itself.
-    private NavOption? CurrentOption() => RunnerPageInstance.GetValue(_expression) as NavOption;
-
-    /// <summary>
-    /// Convert the string a test wrote into the NavValue the binding actually holds.
-    /// AL's TestPage SetValue is string-typed for every control, so the target type has to
-    /// come from the binding, not from the caller — writing a NavText into an Option
-    /// binding throws deep inside the page's own generated setter
-    /// ("Unable to cast object of type 'NavText' to type 'NavOption'"), which says nothing
-    /// about the value that was wrong. A Boolean binding has the same shape of problem
-    /// (#1837): a NavText written into it throws "The input string '...' was not in a
-    /// correct format" instead of setting the field, so Boolean gets the same NavOption-style
-    /// special case — see <see cref="TestPageBooleanValue"/>.
-    ///
-    /// Code and Date bindings (#2054) are the same shape of bug again. A `Code[20]` global's
-    /// generated setter throws "Unable to cast object of type 'NavText' to type 'NavCode'",
-    /// and a `Date` global's throws the same against 'NavDate' — Integer and Text globals
-    /// round-trip fine only because their generated setters happen to accept a NavText and
-    /// coerce it themselves, which Code's and Date's do not. NavCode carries the field's own
-    /// declared length (`Code[20]`), so the replacement is built against the CURRENT bound
-    /// value's own MaxLength rather than a guessed constant.
-    /// </summary>
-    private NavValue ToBoundValue(string value)
-        => RunnerPageInstance.GetValue(_expression) switch
-        {
-            NavOption option => TestPageOptionValue.Resolve(option, value, _page.TryGetOptionCaptions(_controlId, option),
-                $"TestPage SetValue (control {_controlId})"),
-            NavBoolean => TestPageBooleanValue.Resolve(value, Caption),
-            NavCode current => new NavCode(current.MaxLength, value),
-            // #3384: NavDate was the only temporal arm here and accepted only the round-trip
-            // spelling, throwing out-of-scope for text BC reads happily. NavDateTime and NavTime
-            // had no arm at all, so a typed argument reached the page's own generated setter as
-            // a NavText and threw InvalidCastException.
-            //
-            // A DECLINE has to be a typed refusal here, unlike the Rec-bound side. There, the
-            // NavText fall-through reaches BC's own ALValidateAsync and BC raises the refusal
-            // naming the value. A page variable has no validate behind it: the NavText goes
-            // straight into the page's generated setter and comes back out as
-            // "Unable to cast object of type 'NavText' to type 'NavDate'" from inside
-            // NavFormSourceExpression — which names neither the control nor the value the test
-            // wrote. #2054's NavDate arm did raise a typed refusal, and dropping to a bare cast
-            // failure would have been a regression against it.
-            NavDate or NavDateTime or NavTime =>
-                TestPageTemporalValue.TryResolve(FieldType, value, out var temporal)
-                    ? temporal!
-                    : throw new AlRunner.Infrastructure.RunnerOutOfScopeException(
-                        $"TestPage SetValue (control {_controlId})",
-                        $"testpage-temporal-value — '{value}' is not a {FieldType} this BC build "
-                        + "can evaluate, and a page-variable control has no field validate "
-                        + "behind it to refuse the value itself."),
-            _ => ALCompiler.ToNavValue(value),
-        };
-
-    public string Name => Caption;
-    public string Caption => _expression.GetType()
-        .GetProperty("Name", System.Reflection.BindingFlags.Public | System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance)
-        ?.GetValue(_expression) as string ?? string.Empty;
-
-    // The real underlying NavType, not a constant. NavTestField.ALSetValue — the precompiled BC
-    // method the AL compiler emits for every SetValue(<Boolean>) call on this control — asks
-    // THIS property to pick a NavValueMetadata before converting the incoming value to a string
-    // via ITestField.ValueToString (see TestPageBooleanValue's doc comment for the full chain).
-    // A hardcoded NavType.Text made BC's own dispatch treat every page-variable control as text,
-    // so a Boolean write got coerced through Text metadata into BC's "Yes"/"No" textual spelling
-    // (NOT the "True"/"False" ValueToString itself would have produced) before ever reaching our
-    // Value setter — which is why the var-bound and Rec-bound halves of #1837 threw two DIFFERENT
-    // exceptions for the same SetValue(true) call: they disagreed about what string this control
-    // even claimed to receive. A Date global (#2054) failed the SAME way for the SAME reason:
-    // FieldType answering Text sent NavTestField.ALSetValue's DMY2Date(...) argument through
-    // Text metadata instead of Date, and the text it came out as could not be cast back into
-    // the Date binding. Code does not need an entry here — NavCode IS a NavStringValue, so
-    // ALSetValue's own fast path (`value is NavStringValue`) skips FieldType/ValueToString
-    // entirely for it and hands SetValue's literal straight to ToBoundValue above — but it is
-    // listed anyway so a reader checking "does this table cover every case ToBoundValue does"
-    // is not left wondering whether it was missed.
-    public NavType FieldType => RunnerPageInstance.GetValue(_expression) switch
-    {
-        NavOption => NavType.Option,
-        NavBoolean => NavType.Boolean,
-        NavCode => NavType.Code,
-        NavDate => NavType.Date,
-        // #3384: without these two the control claimed Text, so BC's ALSetValue picked Text
-        // metadata for a DateTime/Time argument before ToBoundValue ever saw it.
-        NavDateTime => NavType.DateTime,
-        NavTime => NavType.Time,
-        // #2634/#2534's fix: a Decimal-typed page-global control has to answer NavType.Decimal
-        // here too, the same as an Option/Boolean/Code/Date global already does above -- this
-        // FieldType is what NavTestField.ALSetValue (BC's own precompiled dispatch) uses to pick
-        // a NavValueMetadata before round-tripping through ValueToString, so leaving Decimal out
-        // would have kept a Decimal-typed page variable dispatching as plain Text on the WRITE
-        // side even after the READ side (Value, above) started formatting it correctly.
-        NavDecimal => NavType.Decimal,
-        _ => NavType.Text,
-    };
-    // BC's own NavTestField.CheckError reads all three around every control write, and
-    // NavTestField.ALValidationErrorCount / ALGetValidationError hand the first and fourth
-    // straight to AL. Hardcoded 0/"" made `ValidationErrorCount()` answer 0 after a refusal
-    // real BC counts as 1, and made a refusal escape the setter raw instead of being wrapped
-    // by BC in "Validation error for Field: ..." (#2900). See TestFieldValidationErrors.
-    public int ValidationErrorCount => _validationErrors.Count;
-    public long LastUsedValidationErrorId => _validationErrors.LastUsedId;
-    public long MaxValidationErrorId => _validationErrors.MaxId;
-    public int OptionCount => CurrentOption() is { } option ? TestPageOptionValue.Count(option) : 0;
-
-    // See LiveNavTestField — a control bound to a page variable declares the same properties
-    // as one bound to a record field, and they are read the same way.
-    public bool Enabled  => _page.ControlEnabled(_controlId);
-    public bool Editable => _page.ControlEditable(_controlId);
-    public bool Visible  => _page.ControlVisible(_controlId);
-    public bool HideValue => false;
-    public bool ShowMandatory => false;
-
-    public string GetValidationError(int index) => _validationErrors.Get(index);
-    public void Activate() { }
-    /// <summary>Run the control's OnLookup trigger — see LiveNavTestField.Lookup.</summary>
-    public void Lookup()
-    {
-        var picked = _page.RaiseOnLookup(_controlId, NavText.Create(Value));
-        if (picked != null) Value = picked.ToString();
-    }
-    public void Lookup(NavDataSet dataSet) => Lookup();
-    /// <summary>Run the control's OnAssistEdit trigger — see LiveNavTestField.AssistEdit.</summary>
-    public void AssistEdit() => _page.RaiseOnAssistEdit(_controlId);
-    /// <summary>Run the control's OnDrillDown trigger — see LiveNavTestField.Drilldown.</summary>
-    public void Drilldown() => _page.RaiseOnDrillDown(_controlId);
-    public void Invoke() { }
-
-    // An Option/Enum-bound control renders an ordinal as the text the control SHOWS, the same
-    // spelling the Value getter answers with — issue #2367. BC's own ALAssertEquals/ALSetValue
-    // strip an AL option value down to a bare ordinal before calling this (see
-    // TestPageOptionValue.DisplayOrdinal for the exact chain), so leaving it as
-    // Convert.ToString made AssertEquals compare the ordinal '2' against the control's
-    // 'Pending Approval' and report a mismatch for the value the record actually held.
-    // The Rec-bound sibling above had the identical gap; both are fixed together because both
-    // Value getters already render captions, so either one left alone would keep disagreeing
-    // with its own read side.
-    public string ValueToString(object? value)
-        => TestPageOptionValue.DisplayOrdinal(CurrentOption(), value,
-               CurrentOption() is { } option ? _page.TryGetOptionCaptions(_controlId, option) : null)
-           // #2795: the page-global half of the same rule — see the Rec-bound sibling above.
-           ?? TestPageBooleanValue.FormatObject(value)
-           // #2361: the page-global half of the blank-temporal rule.
-           ?? TestPageBlankTemporalValue.FormatObject(value)
-           ?? Convert.ToString(value, CultureInfo.InvariantCulture) ?? string.Empty;
-    public string GetOption(int index)
-        => CurrentOption() is { } option
-            ? TestPageOptionValue.MemberAt(option, index,
-                _page.TryGetOptionCaptions(_controlId, option))
-            : string.Empty;
-}
-
-/// <summary>Minimal ITestField implementation — all reads return safe defaults.</summary>
-internal sealed class MockITestField : ITestField
-{
-    private string _value = string.Empty;
-
-    public string Value         { get => _value; set => _value = value; }
-    public string Name          => string.Empty;
-    public string Caption       => string.Empty;
-    public NavType FieldType    => NavType.Text;
-    // 0/0/0 and the empty GetValidationError below are FAITHFUL here, not the #2900/#3009
-    // hardcoding that was fixed on the live controls — stated rather than left bare, because
-    // they read identically. This class is only ever built by MockITestPage.GetField, the
-    // degraded path for a page with no live RunnerPageInstance (see
-    // TestPageClientConstructionRule, which warns on stderr when a handle site lands here).
-    // Its Value setter is a plain string store that runs no AL: no OnValidate, no record
-    // write, nothing that can refuse. So there is genuinely never anything to count, and a
-    // ledger here would only ever report the same 0 with more machinery.
-    public int    ValidationErrorCount        => 0;
-    public long   LastUsedValidationErrorId   => 0;
-    public long   MaxValidationErrorId        => 0;
-    public object? ObjectValue               => _value;
-    public int    OptionCount                => 0;
-    public bool   Enabled                   => true;
-    public bool   Editable                  => true;
-    public bool   Visible                   => true;
-    public bool   HideValue                 => false;
-    public bool   ShowMandatory             => false;
-
-    public string GetValidationError(int index)   => string.Empty;
-    public void   Activate()                      { }
-
-    public void   Lookup()                        { }
-    public void   Lookup(NavDataSet dataSet)      { }
-    public void   AssistEdit()                    { }
-    public void   Drilldown()                     { }
-    public void   Invoke()                        { }
-    public string ValueToString(object? value)    => value?.ToString() ?? string.Empty;
-    public string GetOption(int index)            => string.Empty;
-}
-
-/// <summary>Minimal ITestAction implementation — Invoke is a no-op.</summary>
-internal sealed class MockITestAction : ITestAction
-{
-    public void Invoke()         { }
-    public bool Visible          => true;
-    public bool Enabled          => true;
-}
-
-/// <summary>
-/// Dispatches an action against a pageextension's own OnAction trigger when there is no
-/// live RunnerPageInstance for the base page to route LiveNavTestAction through (issue
-/// #1923 — see RunnerPageInstance.TryRaiseExtensionOnlyAction's remarks for why that
-/// happens and what it can and cannot faithfully do). Falls back to a silent no-op, exactly
-/// matching MockITestAction, when no compiled pageextension actually owns this action id —
-/// an id belonging to the (unbuildable) precompiled base page itself, the pre-existing,
-/// narrower gap this deliberately leaves alone rather than expanding scope.
-/// </summary>
-internal sealed class ExtensionOnlyTestAction : ITestAction
-{
-    private readonly LiveNavTestPage _testPage;
-    private readonly object _owner;
-    private readonly NavRecord _record;
-    private readonly int _pageId;
-    private readonly int _actionId;
-
-    public ExtensionOnlyTestAction(LiveNavTestPage testPage, object owner, NavRecord record, int pageId, int actionId)
-    {
-        _testPage = testPage;
-        _owner = owner;
-        _record = record;
-        _pageId = pageId;
-        _actionId = actionId;
-    }
-
-    public void Invoke()
-    {
-        _testPage.SaveCurrentRow();
-        RunnerPageInstance.TryRaiseExtensionOnlyAction(_owner, _record, _pageId, _actionId);
-    }
-
-    public bool Visible => true;
-    public bool Enabled => true;
-}
-
-/// <summary>
-/// A page action driven live: Invoke() runs the page's own OnAction trigger, on the page
-/// instance the TestPage is driving, so the trigger sees the current row.
-///
-/// Visible/Enabled come from the action's own declared properties, which are constants or
-/// expressions evaluated against the page's live state — so an action gated on the current
-/// row (<c>Enabled = RowEditable</c>) reports differently as the cursor moves.
-/// </summary>
-internal sealed class LiveNavTestAction : ITestAction
-{
-    private readonly LiveNavTestPage _testPage;
-    private readonly RunnerPageInstance _page;
-    private readonly int _actionId;
-
-    public LiveNavTestAction(LiveNavTestPage testPage, RunnerPageInstance page, int actionId)
-    {
-        _testPage = testPage;
-        _page = page;
-        _actionId = actionId;
-    }
-
-    /// <summary>
-    /// Save the current row, then run OnAction — BC's order, and the order matters. A real
-    /// client sends the row it is on to the server before it invokes an action, so the AL in
-    /// OnAction reads a <c>Rec</c> that exists in the table, with its AutoSplitKey field
-    /// assigned. Dispatching straight to the trigger let the action see the field values a
-    /// test had just set while the row itself was still nowhere.
-    /// </summary>
-    public void Invoke()
-    {
-        _testPage.SaveCurrentRow();
-
-        // A DISABLED action's Invoke() does nothing, and does not raise. BC's own
-        // Microsoft.Dynamics.Framework.UI.ActionControl.Invoke opens with
-        // `if (!Enabled) return null;`, and TestActionProxy.Invoke — the ITestAction a real
-        // tier hands NavTestAction — carries no gate of its own, so the refusal is silent all
-        // the way up to AL. Measured on BC 28.4: invoking an action declared Enabled = false
-        // leaves the OnAction trigger unrun, raises nothing, and leaves the page usable (the
-        // next action still runs). Corpus codeunit 60583 "TPAR Tests" pins it, for a literal
-        // Enabled = false and for an Enabled bound to an expression that is currently false.
-        //
-        // Running the trigger anyway was a silent wrong answer: AL got an effect it could not
-        // have obtained on a real tier, so a test asserting the effect passed here and failed
-        // there. Observably equivalent to BC now, with one honest difference — BC says nothing
-        // and this says so on [warn], because the shape it produces (an assertion failing one
-        // step later about a missing effect) is exactly what a reader cannot diagnose.
-        //
-        // SaveCurrentRow stays ahead of the gate: BC's TestActionProxy.Invoke calls
-        // parent.ActivateControl(this) before actionControl.Invoke(), so the row reaches the
-        // server whether or not the trigger fires. Not separately measured.
-        if (!_page.ActionEnabled(_actionId))
-        {
-            Console.Error.WriteLine(
-                $"[warn] TestPage: action {_actionId} is not Enabled, so Invoke() did not run "
-                + "its OnAction trigger — real BC refuses a disabled action silently "
-                + "(ActionControl.Invoke: `if (!Enabled) return null;`)");
-            return;
-        }
-
-        _page.RaiseOnAction(_actionId);
-    }
-
-    public bool Visible => _page.ActionVisible(_actionId);
-    public bool Enabled => _page.ActionEnabled(_actionId);
-}
-
-/// <summary>
-/// Minimal ITestPart implementation.
-/// ITestPart extends ITestPage + ITestFilter + IDisposable, so this derives
-/// from MockITestPage which already implements all required members.
-/// </summary>
-internal sealed class MockITestPart : MockITestPage, ITestPart
-{
-    public bool Enabled => true;
-    public bool Visible => true;
-}
-
-/// <summary>
-/// One entry of a part's SubPageLink, in the shape BC's compiler writes into
-/// <c>InfopartPageDefinition.SubFormLink</c>: the PART's field it constrains, the kind, and
-/// either the PARENT's field number (FIELD) or the compiled literal / filter expression
-/// (CONST / FILTER) — see <c>MockTestPage.SubPageLinks</c> for the representation.
-/// </summary>
-internal readonly record struct SubPageLinkEntry(
-    int PartFieldNo, Microsoft.Dynamics.Nav.Types.Metadata.FilterType Kind, int ParentFieldNo, string Value);
-
-/// <summary>
-/// A subpage part driven live: its own page over its own source table, showing only the
-/// rows the SubPageLink selects for the parent's CURRENT row.
-///
-/// The link is re-applied before every operation rather than once at construction, because
-/// NavTestPageBase caches parts for the life of the page: a filter applied once would go
-/// stale the moment the AL test moved the parent to another row, and the part would then
-/// show the previous row's children — a wrong answer that no assertion in the part itself
-/// could distinguish from a right one.
-/// </summary>
-internal sealed class LiveNavTestPart : LiveNavTestPage, ITestPart
-{
-    // Null only when _links has no FIELD entry (issue #2053: a part on a SourceTable-less host
-    // has no parent record and needs none; a CONST/FILTER-only part never reads one either) —
-    // every read below sits inside a FIELD case, so a null parent is never dereferenced.
-    private readonly NavRecord? _parentRecord;
-    private readonly SubPageLinkEntry[] _links;
-
-    /// <param name="record">The part page's own source-table cursor, or null when the part
-    /// page declares NO SourceTable (issue #2195) — a CardPart bound to page globals, the
-    /// info-box shape. Nothing in THIS class needs it in that case, and the reason is a
-    /// property of SubPageLink rather than an observation about the parts seen so far: the
-    /// only behaviour this class adds over LiveNavTestPage is the link, every SubPageLink
-    /// entry names a field of the part's OWN source table, so a part with no source table
-    /// cannot express one and <see cref="_links"/> is necessarily empty. Everything else is
-    /// the base class's null-record path, where each Rec-dependent member refuses by name.</param>
-    /// <param name="parentRecord">The host's current-row cursor; required only when
-    /// <paramref name="links"/> carries a FIELD entry (<see cref="AnyFieldLink"/>).</param>
-    public LiveNavTestPart(NavRecord? record, IReadOnlyDictionary<int, int> controlIdToFieldNo, bool creatable,
-        RunnerPageInstance? page, object owner, int pageId,
-        NavRecord? parentRecord, SubPageLinkEntry[] links)
-        : base(record, controlIdToFieldNo, creatable, page, owner, pageId)
-    {
-        _parentRecord = parentRecord;
-        _links = links;
-    }
-
-    // Constant true, and -- since #3313 -- for a stated reason rather than as an unexplained
-    // hardcode. NavTestPart.ALVisible/ALEnabled read exactly these two properties
-    // (`return testPart.Visible` / `return testPart.Enabled` in Ncl.dll), so hardcoding them
-    // would normally be exactly the silent answer .claude/rules/loud-failures.md refuses.
-    // What makes it faithful is the reachability rule LiveNavTestPage.GetPart now enforces: a
-    // part control whose Visible is the compile-time literal false is not in the test page's
-    // control tree at all, so no LiveNavTestPart is ever constructed for one and AL has no
-    // handle on which to call either accessor. Every part that reaches this class is therefore
-    // one BC would render, for which both answers are true.
-    //
-    // Measured on a real service tier by corpus codeunit 60346
-    // (StefanMaron/BusinessCentral.AL.Language.Tests#227): its first revision tried to assert
-    // the false arm of each pair, and all 8 cloud legs answered "The part with ID = ... was
-    // not found on the page." instead -- Visible() and Enabled() can never be observed
-    // returning false from AL.
-    //
-    // The limit of that argument, stated so a later reader need not re-derive it: a part whose
-    // Visible is an EXPRESSION currently evaluating false IS reachable (expressions are never
-    // compile-time eliminated -- see ControlIsCompileTimeEliminated), and this would answer
-    // true for it. No service tier has measured that shape, because the corpus fixture
-    // declares the literal, so it is not guessed at here.
-    public bool Enabled => true;
-    public bool Visible => true;
-
-    /// <summary>Whether any entry is a FIELD link — the only kind that reads the parent's row.</summary>
-    internal static bool AnyFieldLink(SubPageLinkEntry[] links)
-    {
-        foreach (var link in links)
-            if (link.Kind == Microsoft.Dynamics.Nav.Types.Metadata.FilterType.FIELD) return true;
-        return false;
-    }
-
-    /// <summary>Filter the part's rowset to what its SubPageLink selects for the parent's
-    /// current row: a FIELD entry to the parent's current value, a CONST entry to its literal,
-    /// a FILTER entry to its expression (issue #2469).</summary>
-    private void ApplyLink()
-    {
-        // Nothing to apply, and nothing to demand: an unlinked part shows its own table's
-        // full rowset. The early return is what lets a part page with NO SourceTable exist
-        // at all (issue #2195) — such a part cannot carry any SubPageLink, so it always
-        // lands here, and without the return the RequireRecord below would refuse EVERY
-        // cursor move on it naming "subpage link", a link the part does not have. The move
-        // itself still refuses by its own name through the base class when the AL genuinely
-        // asks a record-less part to navigate.
-        if (_links.Length == 0) return;
-
-        // Past this point the part is linked, which is only expressible against the part's
-        // own source table — so it has one, and this is a guaranteed hit used for its record
-        // rather than for its refusal.
-        var record = RequireRecord("subpage link");
-        foreach (var link in _links)
-        {
-            switch (link.Kind)
-            {
-                case Microsoft.Dynamics.Nav.Types.Metadata.FilterType.FIELD:
-                    record.ALSetRange(link.PartFieldNo, _parentRecord!.GetFieldValue(link.ParentFieldNo));
-                    break;
-                case Microsoft.Dynamics.Nav.Types.Metadata.FilterType.CONST:
-                    record.ALSetFilter(link.PartFieldNo, ConstFilterExpression(record, link.PartFieldNo, link.Value));
-                    break;
-                case Microsoft.Dynamics.Nav.Types.Metadata.FilterType.FILTER:
-                    // Already in BC's filter grammar (the compiler wrote option members as
-                    // ordinals; DependencyPageMetadataXml re-quoted AL identifiers) — BC's own
-                    // filter parser, the one SetFilter uses, reads it. A malformed expression
-                    // raises BC's own NavInvalidFilterExpressionException naming the text.
-                    record.ALSetFilter(link.PartFieldNo, link.Value);
-                    break;
-            }
-        }
-    }
-
-    /// <summary>
-    /// The compiled CONST literal as a filter expression BC's own filter parser reads as that
-    /// ONE value. A Text/Code field gets the literal quoted: the compiler writes
-    /// <c>const('SPECIAL')</c> as the bare text <c>SPECIAL</c>, and a bare literal is parsed as
-    /// an EXPRESSION — a value containing <c>|</c>, <c>..</c>, <c>(</c> or <c>&amp;</c> would be
-    /// read as operators, and an EMPTY literal would clear the filter (SetFilter's own rule for
-    /// <c>''</c>) instead of selecting the blank value. Every other type — an option ordinal,
-    /// a number, a boolean, a date — is handed over as written, exactly the text an AL
-    /// SetFilter call would pass. A field the part's table does not declare is left to
-    /// SetFilter, which refuses it with BC's own error naming the field number.
-    /// </summary>
-    internal static string ConstFilterExpression(NavRecord record, int fieldNo, string value)
-    {
-        var navType = record.MetaTable.TryGetFieldByNo(fieldNo, out var field) ? field.FieldNavType : (NavType?)null;
-        var quote = value.Length == 0 || navType is NavType.Text or NavType.Code;
-        return quote ? "'" + value.Replace("'", "''") + "'" : value;
-    }
-
-    public override bool MoveFirst() { ApplyLink(); return base.MoveFirst(); }
-    public override bool MoveLast() { ApplyLink(); return base.MoveLast(); }
-    public override bool MoveNext() { ApplyLink(); return base.MoveNext(); }
-    public override bool MovePrevious() { ApplyLink(); return base.MovePrevious(); }
-
-    /// <summary>True when this part carries a FIELD SubPageLink — i.e. its rowset depends on
-    /// the PARENT's current row. This is the signal <see cref="LiveNavTestPage.Loaded"/> uses
-    /// to decide whether a parent row-load should refresh this part too (issue #2677). A part
-    /// with no link, or with only CONST/FILTER links, shows a rowset independent of the
-    /// parent's cursor and is never re-positioned by a parent's cursor move; its own initial
-    /// row-load still happens once, from GetPart.</summary>
-    internal bool HasLinks => AnyFieldLink(_links);
-
-    /// <summary>
-    /// Position this part on the row matching its SubPageLink and, if one exists, run its
-    /// OnAfterGetRecord/OnAfterGetCurrRecord — the row-load a real BC FactBox/subpage part
-    /// gets automatically, both when its host opens AND every time the host's own cursor
-    /// moves to a different row. Issue #2677, corpus PR
-    /// StefanMaron/BusinessCentral.AL.Language.Tests#141 (8 BC legs, OBS probes measuring a
-    /// SubPageLink-bound CardPart): with NOTHING ever touching <c>CurrPage.&lt;part&gt;</c> or
-    /// <c>TestPage.&lt;part&gt;</c>, opening the host alone produces
-    /// <c>HostOpen;PartOpen;HostAGCR;PartAGCR</c> — the part's OnOpenPage runs right after the
-    /// host's, and its OnAfterGetRecord/OnAfterGetCurrRecord runs right after the host's own,
-    /// entirely unprompted. A later touch adds nothing (already loaded). Navigating the HOST
-    /// to a different row (GoToRecord) re-fires the part's trigger for the NEW row and does
-    /// NOT re-fire it for the row just left. Before this fix nothing EVER positioned a part's
-    /// own cursor at all — <c>TestPageFactory.TryBuild</c> hands back a BLANK, unfetched
-    /// record, and only an explicit MoveXxx/GoToBookmark call on the PART ITSELF (which
-    /// nothing makes on its behalf) ever reached <see cref="LiveNavTestPage.Loaded"/> — so a
-    /// part whose entire per-row state comes from that trigger (the common FactBox-summary
-    /// shape) stayed at its field defaults for the page's whole life.
-    ///
-    /// Deliberately reuses <c>Loaded(bool)</c> rather than <c>MoveFirst()</c>: MoveFirst()
-    /// also flushes pending parts/rows — a state change appropriate to an AL-driven cursor
-    /// move, not to a parent row simply becoming current. Only the two steps a parent move
-    /// really does are taken: run the FOUND case's trigger, or park on the draft line.
-    ///
-    /// THE NOT-FOUND CASE IS NOT "SHOW NOTHING" (#2923). It used to be, and that was the
-    /// remaining half of #2392 applied to parts: a client renders an editable, insert-allowed
-    /// repeater with no matching rows as exactly one row — its blank new-row line — and a
-    /// write with no <c>New()</c> and no <c>First()</c> of its own lands there. Corpus
-    /// codeunit 60743 <c>EmptyEditableList_SetValueWithoutNewOrFirst_InsertsARow</c> measured
-    /// that on a real service tier for a standalone page; codeunit 60996
-    /// <c>EmptyLinkedPart_WriteWithoutFirst_ValidateSeesTheLinkedKey</c> measures it for a
-    /// LINKED part, which is the shape Microsoft's own document tests use. Leaving the part
-    /// unpositioned sent that write into a record nothing had ever positioned.
-    ///
-    /// <c>AbandonNewRowLine</c> first, because this method is re-entered on every parent move:
-    /// a draft line the part was parked on belongs to the parent being left, and
-    /// <c>Loaded()</c> would not have cleared it.
-    ///
-    /// NOT once-guarded: every call re-applies the link filter and re-finds, which is exactly
-    /// what makes a GoToRecord-driven refresh work. A repeat call for the SAME still-current
-    /// row (a second control read with no intervening parent move) re-runs
-    /// OnAfterGetRecord/OnAfterGetCurrRecord too — unmeasured against BC for that specific
-    /// case (probe 2 only read a control, which triggers a fresh <c>GetPart</c> lookup that
-    /// the `_parts` cache already short-circuits before reaching here at all, so it never
-    /// re-entered this method a second time for the SAME touch).
-    ///
-    /// A record-less part (<see cref="LiveNavTestPage.Record"/> null, the page-globals-only
-    /// CardPart shape from #2195) has no cursor to position and nothing here to do — its
-    /// OnOpenPage is the only trigger such a part gets.
-    /// </summary>
-    /// <summary>
-    /// Whether a record's buffer holds an actual row rather than the blank one a page has
-    /// before its cursor lands anywhere (#3029).
-    ///
-    /// <para>Read off the PRIMARY KEY, because that is what a position is made of and what
-    /// distinguishes the two states here: measured while opening one card over an empty part,
-    /// the host's position reads <c>Field1=0()</c> during EagerlyBuildParts and
-    /// <c>Field1=0(H1)</c> on every call after its cursor lands. Comparing the position STRING
-    /// against a literal would be reading a display format; comparing the key VALUES against
-    /// their initialised state asks the same question of the data.</para>
-    ///
-    /// <para>A table whose whole primary key legitimately holds init values — an integer key at
-    /// 0, a singleton — answers false here and so keeps the pre-#3029 behaviour on this path,
-    /// which is the safe direction: the guard only ever SUPPRESSES a draft-line entry, so a
-    /// false negative costs nothing that was not already happening.</para>
-    /// </summary>
-    private static bool HasCurrentRow(NavRecord record)
-    {
-        var primaryKey = record.MetaTable?.PrimaryKey;
-        if (primaryKey == null || primaryKey.KeyFieldCount == 0) return true;
-        for (var i = 0; i < primaryKey.KeyFieldCount; i++)
-        {
-            var fieldNo = primaryKey.KeyFieldsList[i].FieldNo;
-            // NavValue's own "is this the type's zero" answer, so Code/Text compare against ''
-            // and Integer/Decimal against 0 without this method knowing which it has.
-            var value = record.GetFieldValue(fieldNo);
-            if (value != null && !value.IsZeroOrEmpty) return true;
-        }
-        return false;
-    }
-
-    // The parent row this part was last positioned for, as a position string, or null when it
-    // has never been positioned. Read at the top of ReloadLinkedRow to tell a re-entry for the
-    // SAME parent row from a genuine parent move — see the comment there (#3029).
-    private string? _lastReloadedForParentPosition;
-
-    internal void ReloadLinkedRow()
-    {
-        if (Record is not { } record) return;
-        ApplyLink();
-
-        // IS THIS A PARENT MOVE, OR THE SAME ROW ARRIVING AGAIN? (#3029)
-        //
-        // This method is deliberately not once-guarded — re-applying the link and re-finding is
-        // what makes a GoToRecord-driven refresh work, and that stays. But "the parent row
-        // changed" and "the host called me again about the row I am already on" are different
-        // events, and only the first ends the draft line's row.
-        //
-        // Opening one card over an empty part reaches here THREE times with the parent never
-        // moving: EagerlyBuildParts -> GetPart, MoveFirstDuringOpen -> Loaded ->
-        // RefreshLinkedParts, and ALGoToRecord -> FindRowFromFieldValues -> RefreshLinkedParts.
-        // Each one used to abandon the draft line and enter it again, and entering it raises
-        // the page's OnNewRecord — so merely opening a card cost three firings for a row the
-        // test had not asked for yet, plus a fourth for its own First(). Measured; see the PR
-        // body for the four stacks.
-        var parentPosition = _parentRecord?.ALGetPosition(useCaptions: false);
-        var sameParentRow = _lastReloadedForParentPosition != null
-                            && _lastReloadedForParentPosition == parentPosition;
-        _lastReloadedForParentPosition = parentPosition;
-
-        // NO PARENT ROW YET, SO NO DRAFT LINE YET (#3029). EagerlyBuildParts runs while the
-        // host is still opening and its own cursor has not landed anywhere — measured, the
-        // parent's position reads `Field1=0()` there, against `Field1=0(H1)` on every later
-        // call. A part linked to a parent row that does not exist is showing nothing, and
-        // parking it on a draft line at that moment raised the page's OnNewRecord for a row no
-        // parent owns.
-        //
-        // Only the ENTER is skipped, not the find or the Loaded() below it: the part's cursor
-        // still has to be positioned, and MoveFirstDuringOpen's own reload — which happens once
-        // the host HAS a row — enters the draft line properly. Suppressing the whole method here
-        // would lose the part's OnAfterGetRecord for a part that does have rows.
-        var parentHasNoRow = _parentRecord != null && !HasCurrentRow(_parentRecord);
-
-        // A genuine parent move ends whatever the part was showing, draft line included. A
-        // re-entry for the same row must NOT, or the draft line's started record is discarded
-        // and started again.
-        if (!sameParentRow) AbandonNewRowLine();
-
-        var found = PageInstance?.RaiseOnFindRecord("-")
-                    ?? record.ALFindFirstAsync(DataError.TrapError).GetAwaiter().GetResult();
-        Loaded(found);
-        if (!found && !parentHasNoRow) EnterNewRowLine(record);
-    }
-
-    public override bool FindRowFromTableFieldValues(int[] fieldNos, object[] values, bool forward)
-    {
-        ApplyLink();
-        return base.FindRowFromTableFieldValues(fieldNos, values, forward);
-    }
-
-    /// <summary>
-    /// Start a new row already carrying the link's values — for the fields BC actually
-    /// carries them onto, which is NOT every linked field.
-    ///
-    /// <c>ApplyLink</c> above has just put every entry on the record as a filter, and
-    /// <c>base.InsertEmptyRow</c> then asks BC's own <c>NavForm.NewRecord</c> to start the row
-    /// (<c>RunnerPageInstance.TryNewRecord</c>), which runs
-    /// <c>RecordImplementation.InitRecordFromFilters</c>. That method — Ncl 28.1,
-    /// <c>InitRecordFromFilters(includeNonPrimaryKeyFields, includeIdenticalFilters,
-    /// includeNonPrimaryKeyFieldsForFilterGroups)</c> — copies a field's filter onto the new
-    /// record only when the filter is <c>FilterExpressionType.Equal</c> (exactly one value)
-    /// AND one of: the field is part of the PRIMARY KEY, the page sets
-    /// <c>PopulateAllFields</c>, or the caller names the filter's group.
-    /// <c>NavForm.NewRecordAsync(bool)</c> passes <c>Array.Empty&lt;int&gt;()</c> for the
-    /// groups, so on a TestPage it comes down to key membership.
-    ///
-    /// This loop therefore applies the same key-membership gate. Without it the runner
-    /// stamped every single-valued link onto the new row regardless of the key, which real BC
-    /// does not do: measured on all 8 BC legs of corpus codeunit 60324 "TSPL Tests", a
-    /// <c>New()</c> through a part linked <c>Kind = const(Attachment)</c>, where Kind is not
-    /// part of the line table's key, produced a row with Kind still at Comment — outside the
-    /// part's own filter, which BC then reported as "The view is filtered, and the entry is
-    /// outside the filter". The corpus pins both directions: the same const on a table whose
-    /// key CONTAINS the field IS stamped.
-    ///
-    /// The loop is not redundant with BC's own step even for the fields it does write. It
-    /// covers the record-only fallback in <c>LiveNavTestPage.InsertEmptyRow</c>, where there
-    /// is no page to ask and BC's filter step never runs; where BC's step did run it writes
-    /// the same values again, which is a no-op. A FIELD entry is read from the parent's
-    /// current row; a CONST/FILTER entry is read back through BC's own range accessors on the
-    /// filter <c>ApplyLink</c> just set, so a single-value filter answers the same typed value
-    /// for min and max (an option ORDINAL arrives as a NavOption, not as the text "1") and a
-    /// multi-value or open-ended one raises BC's own error and stamps nothing — the
-    /// <c>Equal</c> half of the same rule, decided by BC rather than re-derived from text.
-    /// </summary>
-    public override void InsertEmptyRow(bool beforeCurrent)
-    {
-        ApplyLink();
-        base.InsertEmptyRow(beforeCurrent);
-        if (_links.Length == 0) return;
-        var record = RequireRecord("subpage link");
-        var primaryKeyFieldNos = PrimaryKeyFieldNos(record);
-        // What was actually stamped, in stamping order — BC's own
-        // `fieldsInitializedFromFilters`, which is the exact set its validate step runs over.
-        var stamped = new List<(int FieldNo, NavValue Value)>();
-        foreach (var link in _links)
-        {
-            // Not part of the primary key: BC leaves it at its Init() value, so the runner
-            // must too — a stamped value here would put a row inside a filter BC would have
-            // reported as outside it.
-            if (!primaryKeyFieldNos.Contains(link.PartFieldNo)) continue;
-            switch (link.Kind)
-            {
-                case Microsoft.Dynamics.Nav.Types.Metadata.FilterType.FIELD:
-                    var linked = _parentRecord!.GetFieldValue(link.ParentFieldNo);
-                    record.SetFieldValue(link.PartFieldNo, linked);
-                    stamped.Add((link.PartFieldNo, linked));
-                    break;
-                default:
-                    if (TryGetSingleFilterValue(record, link.PartFieldNo, out var single))
-                    {
-                        record.SetFieldValue(link.PartFieldNo, single);
-                        stamped.Add((link.PartFieldNo, single));
-                    }
-                    break;
-            }
-        }
-        ValidateStampedFields(record, stamped);
-    }
-
-    /// <summary>
-    /// Run OnValidate on the fields the link just stamped — <c>NavForm.NewRecordAsync</c>'s
-    /// second step, which the runner did not perform at all (issue #2551, gap 2).
-    ///
-    /// <para>BC's body is two steps in order: copy the link's values onto a freshly reset
-    /// buffer, then
-    /// <c>if (ValidateFieldsInOnNewRecord) SourceTable.ValidateFieldsAsync(fieldsInitializedFromFilters, ...)</c>
-    /// — OnValidate on exactly the fields step 1 copied, and nothing else. The runner performed
-    /// step 1 and stopped, so a field carrying a value from the link arrived RAW: its own
-    /// OnValidate never ran, and anything that trigger derives stayed at its Init() default
-    /// while the field itself already held the linked value. That is a wrong answer rather than
-    /// a missing feature, which is why it is fixed rather than declared out of scope.</para>
-    ///
-    /// <para><c>ValidateFieldsInOnNewRecord</c> is a plain auto-property with no setter anywhere
-    /// in Ncl, so nothing in the decompiled runtime says which way it is set — only a service
-    /// tier can answer it. It is answered: corpus codeunit 60653 "NRB Tests"
-    /// (StefanMaron/BusinessCentral.AL.Language.Tests#150) measured on all eight BC legs that a
-    /// New() through a field(...) link DOES run the stamped field's OnValidate. So the flag is
-    /// set by whatever drives a TestPage's New(), and the runner validates unconditionally here
-    /// rather than modelling a flag nothing it can see ever writes.</para>
-    ///
-    /// <para>Copy-then-validate, not validate-during-copy: BC hands its validate step the whole
-    /// set after the copy loop finishes, so an OnValidate on the first stamped field already
-    /// sees the others in place. Validating inside the loop would show it a half-stamped row.</para>
-    ///
-    /// <para>Deliberately NOT wrapped in the <c>CurrFieldNo</c> assignment that
-    /// <c>ValueControl.SetValue</c> uses (#2705). That one models a PAGE-ORIGINATED write, and
-    /// BC's step here is <c>SourceTable.ValidateFieldsAsync</c> — a record-level call, the same
-    /// shape as <c>Rec.Validate</c>, which real BC leaves CurrFieldNo at 0 for. No corpus test
-    /// pins CurrFieldNo during New(), so this follows the mechanism rather than guessing.</para>
-    ///
-    /// <para>Errors propagate. An OnValidate that refuses the linked value is BC refusing to
-    /// start the row, and swallowing it here would hand the test a row real BC never creates.</para>
-    /// </summary>
-    private static void ValidateStampedFields(NavRecord record, List<(int FieldNo, NavValue Value)> stamped)
-    {
-        foreach (var (fieldNo, value) in stamped)
-            record.ALValidateAsync(fieldNo, value, null).GetAwaiter().GetResult();
-    }
-
-    /// <summary>The field numbers making up the record's primary key — the membership test
-    /// <c>NCLMetaField.FieldIsPartOfPrimaryKey</c> answers inside BC, read off the same
-    /// <c>MetaTable.PrimaryKey</c> the AutoSplitKey path already uses so both agree on the key
-    /// shape. Empty for a table whose key metadata is unavailable, which makes the caller stamp
-    /// nothing rather than stamp on a guess.</summary>
-    private static HashSet<int> PrimaryKeyFieldNos(NavRecord record)
-    {
-        var fieldNos = new HashSet<int>();
-        var primaryKey = record.MetaTable?.PrimaryKey;
-        if (primaryKey == null) return fieldNos;
-        for (var i = 0; i < primaryKey.KeyFieldCount; i++)
-            fieldNos.Add(primaryKey.KeyFieldsList[i].FieldNo);
-        return fieldNos;
-    }
-
 }
